### 2017-01-04 #87 @njtierney
visdat package
- What does this package do? (explain in 50 words or less)

`visdat` visualises R data frames so that you can quickly identify data structure and data types. This makes it easier to "get a look at the data" and visually identify abnormalities with a dataset.

- Paste the full DESCRIPTION file inside  a code block below.

```
Package: visdat
Title: Preliminary Visualisation of Data
Version: 0.0.5.9000
Authors@R: person("Nicholas", "Tierney", email = "nicholas.tierney@gmail.com", role = c("aut", "cre"))
Description: visdat makes it easy to visualise your whole dataset so that you
    can visually identify problems.
RoxygenNote: 5.0.1.9000
    plotly (>= 4.5.6),
    stats
    vdiffr

```

- URL for the package (the development repository, not a stylized html page)

<https://github.com/njtierney/visdat>

R users who want to explore their data, particularly when they first receive it.

In terms of visualising missing data as a heatmap, there are a few other packages that have worked on this. The `mi` package used to have a visualisation method for missing data, `missing.pattern.plot` - however this is no longer present in the latest versions. The `Amelia` package has `missmap`, but the default requires some more work to make the final output easier to read.

The `VIM` package provides visualisations for missing data, for example, the `aggr` function provides a histogram of the missingness present in each variable.

In terms of visualising the types of data in a dataset, the [wakefield](https://github.com/trinker/wakefield) package provides the [`table_heat`](https://github.com/trinker/wakefield#visualizing-column-types) function for visualising column data types.

But what makes visdat different?

 `visdat` adheres to the principle that R packages should try to do one thing, it is a simple package that specialises in visualisation of data frames. Amelia and `mi` focus on multiple imputation and missing data methods. `VIM` focusses on visualising missingness and imputation in data, and the `wakefield` package focusses on creating random, reproducible data.

The functionality in visualising missing data for these packages is not the main focus, and so I argue that because visdat is purely about visualising dataframes, it gives it greater scope to work on just one thing.

    - [X] The package contains a [`paper.md`](http://joss.theoj.org/about#paper_structure) with a high-level description.

It succeeds, but there are some notes.

```
checking R code for possible problems ... NOTE
vis_guess: no visible binding for global variable ‘valueGuess’
  valueGuess
```

It does not yet have an rOpenSci footer image

I have not set up pre-commit hooks to ensure that README.md is always newer than README.Rmd, as I'm not sure what `devtools::use_git_hook` does?

I have not added  `#' @noRd` to internal functions as I think it is still useful to have them documented, but I can change this if need be.

- Maëlle Salmon (@masalmon)
- Jenny Bryan (@jennybc)
- Andrew MacDonald (@aammd)
