### 2017-01-04 545de32
add coverage badge to README
