### 2017-01-04 #87 @noamross
That's fine @seaaan, I'll assign once I get through editor checks.  Could you fill out the form at http://ropensci.org/onboarding/? We've started using that to help keep track of reviewers' various expertises.
