### 2017-01-04 #87 @noamross
### Editor checks:

---

#### Editor comments

Currently seeking additional reviewers.  This package fits well into our visualization/reproducibility categories.

* I note that vignette building fails without a recent version of plotly. I suggest you set minimum version numbers for packages in your DESCRIPTION file.
* Below is `goodpractice::gp()` output, with some small things to be addressed. Test coverage is relatively low, which is understandable given the nature of the package, but I suggest you look into using the [*vdiffr*](https://github.com/lionel-/vdiffr) testthat extension to create tests for expected visual output.

```
── GP visdat ─────────

It is good practice to

    R/compare_print.R:10:NA
    R/compare_print.R:11:NA
    R/compare_print.R:12:NA
    R/compare_print.R:13:NA
    R/fingerprint.R:11:NA
    ... and 287 more lines

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80 characters wide. Try make your lines shorter than 80 characters

    R/vis_compare.R:24:1
    R/vis_compare.R:27:1
    R/vis_compare.R:30:1
    R/vis_compare.R:33:1
    R/vis_dat.R:29:1
    ... and 3 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and 1:NCOL(...) expressions. They are error prone and result 1:0 if the expression on the right hand side is zero. Use seq_len() or seq_along() instead.

    R/vis_compare.R:46:26
    R/vis_dat.R:50:26
    R/vis_guess.R:42:26
    R/vis_miss.R:43:24
    R/vis_miss.R:74:26

  ✖ fix this R CMD check NOTE: vis_guess: no visible binding for global variable ‘valueGuess’ Undefined global functions or variables: valueGuess
```
