### 2017-01-05 #87 @noamross
Fine to submit to JOSS. To do so, check off the the JOSS publication options above, and write a `paper.md` file per [JOSS requirements](http://joss.theoj.org/about#author_guidelines). Let us know when you have so reviewers can do the whole review at once.  You can save deposition in a repository for a DOI for after you have made changes post-review.
