### 2017-01-05 #87 @noamross
Oh, and the comments I put above, e.g., testing, can be addressed once all other reviews are in, too.  If they were major I would have held up assigning reviewers until they were addressed.
