+++ man/vis_dat_ly.Rd
@@ -25,4 +25,3 @@ vis_dat_ly(airquality)
 
 
 }
-
