+++ man/fingerprint.Rd
@@ -12,4 +12,3 @@ fingerprint(x)
 \description{
 \code{fingerprint} is an internal function that takes the "fingerprint" of a dataframe, and (currently) replaces the contents (x) with the class of a given object, unless it is missing (coded as NA), in which case it leaves it as NA. The name "fingerprint" is taken from the csv-fingerprint, of which this package is based.
 }
-
