### 2017-01-07 1878df4
add flip=TRUE argument to visdat and vismiss as in #35
