+++ R/vis_miss.R
@@ -10,6 +10,8 @@
 #'
 #' @param show_perc logical TRUE/FALSE. TRUE now adds in the \% of missing/complete data in the whole dataset into the legend. Default value is TRUE
 #'
+#' @param flip logical if TRUE, will flip the axis labels to be on top, resembling a dataframe
+#'
 #' @examples
 #'
 #' library(visdat)
@@ -19,11 +21,16 @@
 #' vis_miss(airquality, cluster = TRUE)
 #'
 #' vis_miss(airquality, sort_miss = TRUE)
+#'
+#' # flip the axis to look more like a dataframe
+#' vis_miss(airquality, flip = TRUE)
+#'
 #' @export
 vis_miss <- function(x,
                      cluster = FALSE,
                      sort_miss = FALSE,
-                     show_perc = TRUE){
+                     show_perc = TRUE,
+                     flip = FALSE){
   # make a TRUE/FALSE matrix of the data.
   # This tells us whether it is missing (true) or not (false)
   # x = airquality
@@ -102,6 +109,7 @@ vis_miss <- function(x,
   }
 
     # then we plot it
+  vis_miss_plot <-
     ggplot2::ggplot(data = d,
            ggplot2::aes_string(x = "variables",
                       y = "rows",
@@ -130,5 +138,18 @@ vis_miss <- function(x,
       # guides(fill = guide_legend(title = "Type"))
   # Thanks to http://www.markhneedham.com/blog/2015/02/27/rggplot-controlling-x-axis-order/
   # For the tip on using scale_x_discrete
+
+  if(flip == FALSE){
+    return(vis_miss_plot)
+  } else if(flip == TRUE){
+      suppressMessages({
+        vis_miss_plot +
+          ggplot2::scale_y_reverse() +
+          ggplot2::scale_x_discrete(position = "top",
+                                    limits = col_order_index) +
+          ggplot2::theme(axis.text.x = ggplot2::element_text(hjust = 0.5)) +
+          ggplot2::labs(x = "")
+      })
+  }
 
 } # end of function
