+++ man/visdat.Rd
@@ -8,4 +8,3 @@
 \description{
 visdat
 }
-
