### 2017-01-07 9b2556e
version bump and add flip to news; add minimum version for plotly
