+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: visdat
 Title: Preliminary Visualisation of Data
-Version: 0.0.4.9990
+Version: 0.0.4.9999
 Authors@R: person("Nicholas", "Tierney", email = "nicholas.tierney@gmail.com", role = c("aut", "cre"))
 Description: visdat makes it easy to visualise your whole dataset so that you
     can visually identify problems.
@@ -8,14 +8,14 @@ Depends:
     R (>= 3.2.2)
 License: MIT + file LICENSE
 LazyData: true
-RoxygenNote: 5.0.1
+RoxygenNote: 5.0.1.9000
 Imports:
     ggplot2,
     tidyr,
     dplyr,
     purrr,
     readr,
-    plotly,
+    plotly (>= 4.5.6),
     magrittr,
     stats
 URL: https://njtierney.com/visdat
