### 2017-01-07 4d2f22c
add paper folder for JOSS
