### 2017-01-07 77a14e5
Clean up typos in README.Rmd.
