+++ README.Rmd
@@ -38,17 +38,12 @@ Initially inspired by [`csv-fingerprint`](https://github.com/setosa/csv-fingerpr
 The name `visdat` was chosen as I think in the future it could be integrated with [`testdat`](https://github.com/ropensci/testdat). The idea being that first you visualise your data (`visdat`), then you run tests from `testdat` to fix them.
 
 There are two main commands in the `visdat` package: 
-- `vis_dat()` 
-- `vis_miss()`
 
 - `vis_dat()` visualises a dataframe showing you what the classes of the columns are, and also displaying the missing data.
 
 - `vis_miss()` visualises just the missing data, and allows for missingness to be clustered and columns rearranged. `vis_miss()` is similar to `missing.pattern.plot` from the `mi` package. Unfortunately `missing.pattern.plot` is no longer in the `mi` package (well, as of 14/02/2016).
 
-There are two some experimental functions:
-
-- `vis_guess()`
-- `vis_compare()`
+There are two experimental functions:
 
 -  `vis_guess()` has a guess at what the value of each cell, using `readr::parse_guess`. This means that "10.1" will return "double", and `10.1` will return "double", and `01/01/01` will return "date". Keep in mind that it is a **guess** at what each cell is, so you can't trust this fully.
 
@@ -70,7 +65,7 @@ vis_dat(airquality)
 
 The classes are represented on the legend, and missing data represented by grey. 
 
-by default, `vis_dat` sorts the columns according to the type of the data in the vectors. You can turn this off by setting `sort_type == FALSE`. This feature is better illustrated using the `airquality` dataset from base R.
+by default, `vis_dat` sorts the columns according to the type of the data in the vectors. You can turn this off by setting `sort_type = FALSE`. This feature is better illustrated using the `airquality` dataset from base R.
 
 ```{r vis-dat-aq-sort-type}
 
@@ -79,11 +74,11 @@ vis_dat(airquality,
 
 ```
 
-The plot above tells us that R reads this dataset as having characters, and integer values, along with some missing data in `Ozone` and `Solar.R`.
+The plot above tells us that R reads this dataset as having characters and integer values, along with some missing data in `Ozone` and `Solar.R`.
 
-### with different data
+### With many kinds of data
 
-To demonstrate what visdat looks like when you have different kinds of data, we can look at the `typical_data` from within visdat, created using the excellent [`wakefield package`](https://github.com/trinker/wakefield).
+To demonstrate what visdat looks like when you have different kinds of data, we can use `visdat` to look at the dataset `typical_data`, which is from the excellent [`wakefield package`](https://github.com/trinker/wakefield).
 
 ```{r vis-dat-typical-data}
 
@@ -100,7 +95,7 @@ vis_dat(typical_larger_data)
 ```
 
 
-## using `vis_miss()`
+## Using `vis_miss()`
 
 We can explore the missing data further using `vis_miss()`
 
@@ -152,7 +147,7 @@ vis_miss(mtcars)
 
 # Experimental features
 
-## using `vis_compare()`
+## Using `vis_compare()`
 
 Sometimes you want to see what has changed in your data. `vis_compare()` helps with that. It is currently only just barely working, so keep in mind that this is very much in its beta stages.
 
@@ -167,7 +162,7 @@ vis_compare(iris_diff, iris)
 
 ```
 
-Here the differences are marked in red.
+Here the differences are marked in blue.
 
 If you try and compare differences when the dimensions are different, you get an ugly error.
 
@@ -180,7 +175,7 @@ vis_compare(iris, iris_diff_2)
[…]
