### 2017-01-08 17a27a7
Correct typos in vignettes.
