### 2017-01-08 e2533fd
ensure that all line lengths < 80 characters
