### 2017-01-08 df5e637
replace 1:nrow(.) with seq_len(nrow(.))
