+++ man/vis_dat.Rd
@@ -13,7 +13,7 @@ vis_dat(x, sort_type = TRUE, palette = "default", flip = FALSE)
 
 \item{palette}{character "default", "qual" or "cb_safe". "default" (the default) provides the stock ggplot scale for separating the colours. "qual" uses an experimental qualitative colour scheme for providing distinct colours for each Type. "cb_safe" is a set of colours that are appropriate for those with colourblindness. "qual" and "cb_safe" are drawn from http://colorbrewer2.org/.}
 
-\item{flip}{logical if TRUE, will flip the axis labels to be on top, resembling a dataframe}
+\item{flip}{logical TRUE, will flip the axis labels to be on top, resembling a dataframe}
 }
 \description{
 \code{vis_dat} gives you an at-a-glance ggplot object of what is inside a dataframe. Cells are coloured according to what class they are and whether the values are missing. As \code{vis_dat} returns a ggplot object, it is very easy to customize and change labels, and customize the plot
