### 2017-01-08 8c4df64
update NEWS
