### 2017-01-08 c5cdc1c
configure for vdiffr
