### 2017-01-08 56df1b3
add tests for vis_dat and vis_miss
