### 2017-01-08 9d22f4b
add vdiffr to Suggests and minor change to vis_guess
