### 2017-01-08 7a94d91
change warnings to messages
