### 2017-01-08 bf3a8f2
add vdiffr tests for vis-compare , vis-guess, and vis-miss-ly
