### 2017-01-08 de849d8
remove vis_dat_ly, put into issue #36
