### 2017-01-08 465a2c9
add globalVariables "valueGuess".
This seems to be best practice...ish? https://stackoverflow.com/questions/9439256/how-can-i-handle-r-cmd-check-no-visible-binding-for-global-variable-notes-when
