+++ R/visdat-package.r
@@ -6,3 +6,4 @@
 NULL
 
 if(getRversion() >= "2.15.1")  utils::globalVariables(c("."))
+globalVariables(c("valueGuess"))
