### 2017-01-08 #87 @njtierney
OK great, just working on the JOSS paper now, will hopefully have it done this evening.

Re your comments @noamross

```
I note that vignette building fails without a recent version of plotly. I suggest you set minimum version numbers for packages in your DESCRIPTION file.
```
DESCRIPTION file has been updated to reflect this [here](https://github.com/njtierney/visdat/blob/master/DESCRIPTION)

```
Below is goodpractice::gp() output, with some small things to be addressed. Test coverage is relatively low, which is understandable given the nature of the package, but I suggest you look into using the vdiffr testthat extension to create tests for expected visual output.
```

Here is my updated output from `goodpractice::gp()`

```
── GP visdat ────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions,
    and all package code in general. At
    least some lines are not covered in
    this package.

    R/vis_compare.R:33:NA
    R/vis_compare.R:34:NA

─────────────────────────────────────────────
```

Test coverage is now at [99% using vdiffr](https://codecov.io/github/njtierney/visdat?branch=master). I think I have used `vdiffr` correctly, it looks like the main use is to check that the plot reliably produces the same output, and the test uses `expect_doppelganger()`.

Thank you again for the feedback, I will let you know when I add the paper.md doc.
