### 2017-01-08 #87 @njtierney
Ack, sorry if I jumped the gun replying to the comments.
