### 2017-01-08 cda9514
put test_that around the tests
