+++ tests/testthat/test-vis-guess.R
@@ -3,4 +3,6 @@ context("vis_guess")
 # try out all the options
 vis_guess_plot <- vis_guess(airquality)
 
-vdiffr::expect_doppelganger("vis_guess vanilla", vis_guess_plot)
+test_that("vis_miss_ly creates the right plot",{
+  vdiffr::expect_doppelganger("vis_guess vanilla", vis_guess_plot)
+})
