+++ tests/testthat/test-vis-miss-ly.R
@@ -2,5 +2,7 @@ context("vis_miss_ly")
 
 vis_miss_ly_plot <- vis_miss_ly(airquality)
 
-vdiffr::expect_doppelganger("vis_miss_ly is plotly, hopefully this works!",
+test_that("vis_miss_ly creates the right plot",{
+vdiffr::expect_doppelganger("vis_miss_ly",
                             vis_miss_ly_plot)
+})
