### 2017-01-08 18269b5
update paper notes
