### 2017-01-08 594f612
update the paper, reduce excess content
