+++ paper/paper.Rmd
@@ -12,7 +12,7 @@ affiliations:
  - name: Queensland University of Technology
    index: 1
 date: 09 January 2017
-# bibliography: paper.bib
+bibliography: paper.bib
 output: 
   html_document: 
     keep_md: yes
@@ -20,7 +20,7 @@ output:
 
 # Summary
 
-As described by Hadley Wickham, there are 6 phases of data science:
+As described by Hadley Wickham [@wickham2016], there are 6 phases of data science:
 
 ![Process of Data Science](http://r4ds.had.co.nz/diagrams/data-science.png)
 
@@ -34,7 +34,7 @@ head(iris)
 
 ```
 
-Another alterantive is `dplyr::glimpse`
+Another alterantive is `glimpse`, from the `dplyr` package [@dplyr]
 
 ```{r glimpse}
 
@@ -57,7 +57,7 @@ Looking at this, you might then ask:
 
 And you might start to wonder what else is different, what else changed? 
 
-And it might be a bit unclear where to go from there. Do you plot the data? Why does my plot look wierd? What are these other strange features in the data? The `visdat` package provides visualisations of an entire dataframe at once. Initially inspired by [`csv-fingerprint`](https://github.com/setosa/csv-fingerprint), `visdat` provides tools to create heatmap-like visualisations of an entire dataframe. `visdat` [@Tierney2017] is an R [@Rcore] package provides 2 main functions `vis_dat` and `vis_miss`.
+And it might be a bit unclear where to go from there. Do you plot the data? Why does my plot look wierd? What are these other strange features in the data? The `visdat` package provides visualisations of an entire dataframe at once. Initially inspired by [`csv-fingerprint`](https://github.com/setosa/csv-fingerprint), `visdat` provides tools to create heatmap-like visualisations of an entire dataframe. `visdat` is an R [@Rcore] package provides 2 main functions `vis_dat` and `vis_miss`.
 
 `vis_dat()` helps explore the data class structure and missingness:
 
@@ -75,6 +75,6 @@ vis_miss(typical_data)
 
 ```
 
-These functions provide useful tools to help "get a look at the data", using principled visualisation techniques. The plots are built using ggplot2 [@ggplot2], which provides a consistent and powerful framework for visualisations. Which means that users can customise and extend graphics from visdat very easily.
+These functions provide useful tools to help "get a look at the data", using principled visualisation techniques. The plots are built using ggplot2 [@ggplot2], which provides a consistent and powerful framework for visualisations. This means that users can customise and extend graphics from visdat very easily.
 
 # References
