### 2017-01-08 #87 @njtierney
OK @noamross I have pushed my [paper.md](https://github.com/njtierney/visdat/blob/master/paper/paper.md) into the github repo under the folder `visdat/paper/`.

Thank you again!
