### 2017-01-08 b922cbb
update NEWS + Description for submission to rOpenSci onboarding and JOSS
