### 2017-01-08 #87 @njtierney
I've also updated the DESCRIPTION file as well.
