### 2017-01-09 82a47e8
add yaml to paper.md
this gets lopped off when I go from .Rmd -> .md
