+++ README.md
@@ -55,7 +55,7 @@ library(visdat)
 vis_dat(airquality)
 ```
 
-![](README-vis-dat-aq-1.png)
+![](README-figs/README-vis-dat-aq-1.png)
 
 The classes are represented on the legend, and missing data represented by grey.
 
@@ -67,7 +67,7 @@ vis_dat(airquality,
         sort_type = FALSE)
 ```
 
-![](README-vis-dat-aq-sort-type-1.png)
+![](README-figs/README-vis-dat-aq-sort-type-1.png)
 
 The plot above tells us that R reads this dataset as having characters, and integer values, along with some missing data in `Ozone` and `Solar.R`.
 
@@ -82,7 +82,7 @@ vis_dat(typical_data)
 #> be dropped
 ```
 
-![](README-vis-dat-typical-data-1.png)
+![](README-figs/README-vis-dat-typical-data-1.png)
 
 We can also look into using even wider data, looking at `typical_larger_data`
 
@@ -93,7 +93,7 @@ vis_dat(typical_larger_data)
 #> be dropped
 ```
 
-![](README-vis-dat-typical-larger-data-1.png)
+![](README-figs/README-vis-dat-typical-larger-data-1.png)
 
 using `vis_miss()`
 ------------------
@@ -105,7 +105,7 @@ We can explore the missing data further using `vis_miss()`
 vis_miss(airquality)
 ```
 
-![](README-vis-miss-aq-1.png)
+![](README-figs/README-vis-miss-aq-1.png)
 
 The percentages of missing/complete in `vis_miss` are accurate to 1 decimal place.
 
@@ -117,7 +117,7 @@ vis_miss(airquality,
          cluster = TRUE)
 ```
 
-![](README-vis-miss-aq-cluster-1.png)
+![](README-figs/README-vis-miss-aq-cluster-1.png)
 
 The columns can also just be arranged by columns with most missingness, by setting `sort_miss = TRUE`.
 
@@ -127,7 +127,7 @@ vis_miss(airquality,
          sort_miss = TRUE)
 ```
 
-![](README-vis-miss-aq-sort-miss-1.png)
+![](README-figs/README-vis-miss-aq-sort-miss-1.png)
 
 When there is &lt;0.1% of missingness, `vis_miss` indicates that there is &gt;1% missingness.
 
@@ -142,7 +142,7 @@ vis_miss(test_miss_df)
 #> be dropped
 ```
 
-![](README-vis-miss-test-1.png)
+![](README-figs/README-vis-miss-test-1.png)
 
 `vis_miss` will also indicate when there is no missing data at all
 
@@ -151,7 +151,7 @@ vis_miss(test_miss_df)
 vis_miss(mtcars)
 ```
 
-![](README-vis-miss-mtcars-1.png)
+![](README-figs/README-vis-miss-mtcars-1.png)
 
 Experimental features
 =====================
@@ -180,7 +180,7 @@ vis_compare(iris_diff, iris)
 #> be dropped
 ```
 
-![](README-vis-compare-iris-1.png)
+![](README-figs/README-vis-compare-iris-1.png)
 
 Here the differences are marked in blue.
 
@@ -233,7 +233,7 @@ vis_guess(messy_df)
 #>           post an issue at https://github.com/njtierney/visdat/issues
 ```
 
-![](README-vis-guess-messy-df-1.png)
+![](README-figs/README-vis-guess-messy-df-1.png)
 
 So here we see that there are many different kinds of data in your dataframe. As an analyst this might be a depressing finding. Compare this to `vis_dat`.
 
@@ -242,7 +242,7 @@ So here we see that there are many different kinds of data in your dataframe. As
 vis_dat(messy_df)
 ```
 
[…]
