### 2017-01-14 b3cafe7
A few more wording changes to the README
