### 2017-01-15 #87 @njtierney
Thank you very much @seaaan for the review, and for taking the time to cover it in as so much detail. I agree with you very much that this is a great process set up by rOpenSci.

@noamross and @sckott, is it OK for me to address these comments now, or should I wait for the second reviewer, to avoid doubling up?

I'm really looking forward to addressing these changes, I feel like visdat is going to be much better as a result of this process.
