### 2017-01-15 356431d
update pkgdown with typo fixes from pull #37 by @seaaan
