### 2017-01-16 #87 @noamross
Thank you @seaaan for the thorough review.  @njtierney you can reply here and make changes, but I suggest you use a branch or work locally so that @batpigandme may have a stable version to review.

Two quick notes:

- @seaaan, it looks like @njtierney is use the development version of **roxygen2**, which allows one to write documentation with markdown syntax which is converted to `.Rd` syntax.  This is fine.
- @arfon, can you comment if the [`paper.md`](https://github.com/njtierney/visdat/blob/master/paper/paper.md) isn't in JOSS format. I believe images are allowed, correct?
