### 2017-01-16 #87 @seaaan
Cool, I didn't know that **roxygen2** would do that. That definitely makes it easier. For some reason, however, the markdown syntax didn't get converted to `.Rd` syntax in parts of the description section of [`vis_compare.Rd`](https://github.com/njtierney/visdat/blob/master/man/vis_compare.Rd).
