### 2017-01-16 #87 @arfon
> @arfon, can you comment if the paper.md isn't in JOSS format. I believe images are allowed, correct?

Images are ok. I'm not exactly sure what happens to those ` ```r` code blocks - we'll see when I try and compile it later :-)
