### 2017-01-17 #87 @njtierney
@batpigandme At this stage I probably won't have time to address @seaaan 's comments until the weekend, no need to wait on me, :)

Re markdown in the Rd files, @noamross and @seaaan, I need to one more thing to the DESCRIPTION file to get the markdown syntax to work - [`Roxygen: list(markdown = TRUE)`] (more [here](https://github.com/klutometis/roxygen/blob/master/vignettes/markdown.md))  so Sean's comment makes sense.
