### 2017-01-23 7358a15
Minor edits to README (e.g. punctuation, spelling); changes of more significance noted in package review.
