+++ README.md
@@ -18,7 +18,7 @@ devtools::install_github("njtierney/visdat")
 What does visdat do?
 ====================
 
-Initially inspired by [`csv-fingerprint`](https://github.com/setosa/csv-fingerprint), `vis_dat` helps you visualise a dataframe and "get a look at the data" by displaying the variable classes in a dataframe as a plot with `vis_dat`, and getting a brief look into missing data patterns `vis_miss`.
+Initially inspired by [`csv-fingerprint`](https://github.com/setosa/csv-fingerprint), `vis_dat` helps you visualise a dataframe and "get a look at the data" by displaying the variable classes in a dataframe as a plot with `vis_dat`, and getting a brief look into missing data patterns using `vis_miss`.
 
 The name `visdat` was chosen as I think in the future it could be integrated with [`testdat`](https://github.com/ropensci/testdat). The idea being that first you visualise your data (`visdat`), then you run tests from `testdat` to fix them.
 
@@ -26,7 +26,7 @@ There are two main commands in the `visdat` package:
 
 -   `vis_dat()` visualises a dataframe showing you what the classes of the columns are, and also displaying the missing data.
 
--   `vis_miss()` visualises just the missing data, and allows for missingness to be clustered and columns rearranged. `vis_miss()` is similar to `missing.pattern.plot` from the `mi` package. Unfortunately `missing.pattern.plot` is no longer in the `mi` package (well, as of 14/02/2016).
+-   `vis_miss()` visualises just the missing data, and allows for missingness to be clustered and columns rearranged. `vis_miss()` is similar to `missing.pattern.plot` from the [`mi`](https://cran.r-project.org/web/packages/mi/index.html) package. Unfortunately `missing.pattern.plot` is no longer in the `mi` package (well, as of 14/02/2016).
 
 There are two experimental functions:
 
@@ -40,30 +40,34 @@ Examples
 Using `vis_dat()`
 -----------------
 
-Let's see what's inside the dataset `airquality`
+Let's see what's inside the `airquality` dataset from base R.
 
 ``` r
 
 library(visdat)
 
 vis_dat(airquality)
+#> dmap() is deprecated. Please use the new colwise family in dplyr.
+#> E.g., summarise_all(), mutate_all(), etc.
 ```
 
 ![](README-figs/README-vis-dat-aq-1.png)
 
-The classes are represented on the legend, and missing data represented by grey.
+The classes are represented on the legend, and missing data represented by grey. The column/variable names are listed on the x axis.
 
-by default, `vis_dat` sorts the columns according to the type of the data in the vectors. You can turn this off by setting `sort_type = FALSE`. This feature is better illustrated using the `airquality` dataset from base R.
+By default, `vis_dat` sorts the columns according to the type of the data in the vectors. You can turn this off by setting `sort_type = FALSE`.
 
 ``` r
 
 vis_dat(airquality, 
         sort_type = FALSE)
+#> dmap() is deprecated. Please use the new colwise family in dplyr.
+#> E.g., summarise_all(), mutate_all(), etc.
 ```
 
 ![](README-figs/README-vis-dat-aq-sort-type-1.png)
 
-The plot above tells us that R reads this dataset as having characters and integer values, along with some missing data in `Ozone` and `Solar.R`.
+The plot above tells us that R reads this dataset as having numeric and integer values, with some missing data in `Ozone` and `Solar.R`.
 
 ### With many kinds of data
 
@@ -72,6 +76,8 @@ To demonstrate what visdat looks like when you have different kinds of data, we 
 ``` r
 
 vis_dat(typical_data)
+#> dmap() is deprecated. Please use the new colwise family in dplyr.
[…]
