### 2017-01-24 #87 @batpigandme
>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [X] **A short summary** describing the high-level functionality of the software
>- [X] **Authors:**  A list of authors with their affiliations
>- [X] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

~8
---

`visdat` provides a quick way for the user to visually examine the structure of their data set, and, more specifically, where and what kinds of data are missing. Having now put it to use in a few of my own projects, I can confidently say that it is an incredibly useful early step in the data analysis workflow. `vis_miss`, in particular, is helpful for scoping the task at hand. `vis_guess` brings a great, new level of detail to quick, visual examination of the structure of a dataframe; among other things, I imagine this being useful for instructional purposes. (If I exposited all of the positive feedback in the comments, it would be a tome). `vis_dat`'s power lies in the fact that it's essentially a visual analogue to one's data, if imagined as a frame/spreadsheet. In order to enhance this, I would suggest that the charts, by default, "read" in the spirit of this analogy by going from top to bottom.

As a relative novice when it comes to the inner workings of R packages, several of my comments target fairly minor things (e.g. warning messages) that might require more explanation and/or pose barriers to adoption for the new(ish) R user. Of course, feel free to disregard these at will! (n.b. some of this was written before and some after @seaaan's review, so I've tried to avoid redundancy where possible, please feel free to ask for any clarification or reformatting, as I'm a total package-reviewing neophyte).

##### README
- The `README` section flows nicely, and has illustrative examples, all of which I was able to run.
- I have made some minor changes to the wording in README and submitted them as a pull request [here](https://github.com/njtierney/visdat/pull/38), most of which I think are self-explanatory. However, I'm noting (below) changes which you may want to review more closely.
  - `vis_dat(airquality)` is _numeric_ and integer values, when I run it. Thus, I've swapped numeric in for `character` as it was described.
  - Added canonical link to the [`mi`](https://cran.r-project.org/web/packages/mi/index.html) package, since you do so for `wakefield`
- Though the `README` suggests that when less than one percent is missing, _“`vis_miss` indicates that there is >1%”_, the chart that results from `vis_miss(test_miss_df)` reads _Missing (<0.1%)_ in the legend
- When running the first example from `README`, (`vis_dat(airquality)`), the user will get a warning re. depecation of `dmap()`. If this is _not_ of consequence to the package's functionality, then you might consider adding it to the suppressed messages. If it's something most users supress locally, then you might suggest that somewhere in the README.

```
   #> dmap() is deprecated. Please use the new colwise family in dplyr.
   #> E.g., summarise_all(), mutate_all(), etc.
```
- Minor suggestion (which may or may not work comply with the actual vignette formatting) the purpose of demonstrating the difference between `vis_guess()` and `vis_dat`, you might consider showing the two side by side (perhaps have the `vis_guess()` example alone and then side by side). For this you could just add `fig.show='hold', out.width='50%'` to the chunk options in the Rmd.
- Based on the README, the status of `vis_dat_ly` isn't totally clear (since it's not yet a function of its own).
- I didn't quite understand the **Visualising expectations** section. That said, I've also never used `expectation` in `assertr`— so, this may fall under the category of things that those who need to know do, and, thus, no more need be said.

- Overall these read well, and cover all of the exported functions.
- Some of the same copy edits I made in `README` I would also suggest in the vignettes (e.g. the compare this to...Where sentence that sort of goes around the `vis_dat` figure in the `vis_guess` example section)
- The experimental features vignette is a bit hard to follow at the moment, which may be in part due to the layout (the warning messages take up a good bit of space). The figure is also quite small. I'd recommend adjusting the settings to match your main vignette to make it more legible.
- The experimental features vignette is a bit hard to follow at the moment, which may be in part due to the layout (the warning messages take up a good bit of space). The figure is also quite small. I'd recommend adjusting the settings to match your main vignette to make it more legible.
- `vis_dat_ly()`, and `vis_miss_ly()` aren't in the experimental vignette, though the opening paragraph suggests that they will be

##### Help
- Nice, concise descriptions which, for the most part, reflect the information in `README` and the vignettes.
- The `vis_guess` description is a bit hard to follow. I'd start by stating what it _does_ do, before making the comparison to `vis_dat`, just so it's a bit more self-contained.
- The `visdat` and `visdat-package` sections are empty, when, I assume, you'll want the description to be included (right now description just says "visdat").

##### Examples
- Note: the function examples in R all throw warning: `dmap() is deprecated`.
- As noted in @seaaan's comments, there's also the warning re. dataframes of different dimensions
  - e.g. `example("vis_compare")` results in
```
  #>  Warning messages:
  #> 1: In if (dim(df1) != dim(df2)) { :
  #>  the condition has length > 1 and only the first element will be used
  #> 2: attributes are not identical across measure variables; they will be dropped
  #> 3: attributes are not identical across measure variables; they will be dropped
```

- `vis_compare` example also seems to cluster by default (all different data is shown at the bottom, of the two columns)

##### Community guidelines
- I don't believe these are currently in the package.

##### Paper
- The `vis_miss` chunk is unnamed. I'd put a name in for you, but I can't knit the doc (see below). Either way, quick fix.
- I'm getting an error 83 because `pandoc-citeproc` is missing. I believe this is something I'm missing on my machine, and, thus you don't need to worry about. But, just in case, I'm letting you know!

##### Installation and Tests
Installation built successfully (on three different computers, and versions of R at that), as did automated testing. Good work!

##### Functionality
All of the functions run as expected, and as described. My feedback here primarily concerns messages and warnings, that apply to several functions.
- Warning messages can be scary, so it's always a relief see them acknowledged or shown in vignette and/or README examples (as is done re. non-identical attributes).
- `vis_compare` (a function for which I can imagine plenty of use cases) is currently returning enough text that it may leave the user confused as to what's most important, or an error on their end, etc. Currently, running `vis_compare(iris_diff, iris)` (as in the given example), returns:

```
#> vis_compare is still in BETA! If you have suggestions or errors,
#>        post an issue at https://github.com/njtierney/visdat/issues
#>  dmap() is deprecated. Please use the new colwise family in dplyr.
#>  E.g., summarise_all(), mutate_all(), etc.
#>  Warning messages:
#>  1: In if (dim(df1) != dim(df2)) { :
#>    the condition has length > 1 and only the first element will be used
#>  2: attributes are not identical across measure variables; they will be dropped
#>  3: attributes are not identical across measure variables; they will be dropped
```

  The function runs fine, and appears to have worked, but it's enough "warning" text, that it may warrant concern for the user.
- `vis_guess` is where the cell-to-row visual disparity becomes most confusing. I think this is a fantastic function, _especially_ for beginners, and so the ability to essentially "read" the visual, as one would a table (top to bottom), would be invaluable.
- `vis_dat` with plotly / interactive functions are great, and, again, here it would be even more helpful if one could "read" the charts as one would a table.
