### 2017-01-24 #87 @njtierney
Thank you very much for your thorough and thoughtful review, @batpigandme ! This is great feedback, and I think that visdat will be much improved because of these reviews.

I will make some time this weekend to address these reviewer comments.

Once again, thank you to the reviewers @seaaan and @batpigandme, and the editors @noamross and @sckott for their time. This is a fantastic review process.
