### 2017-01-29 #87 @njtierney
Hi Everyone,

At the moment I need to get a major thesis paper to my supervisors in the next few days, so that is (unfortunately!) taking priority at the moment.

I am still responding to the reviewer comments, if it is OK with the reviewers and rOpenSci, would it be OK if I have my response completed before the end of next weekend, the 4th February?
