+++ paper/response.Rmd
@@ -1,0 +1,410 @@
+# Response to reviewers
+
+I would like to thank the reviewers for taking the time to review visdat so thoroughly. It is fantastic to hear comments such as:
+
+> 
+
+,
+
+>
+
+And
+
+>
+
+...
+
+
+Consider the above my response to the positive feedback in this review. I will now respond to suggsted changes and other constructive criticisms.
+
+# @seaaan's Review
+
+1. Documentation: README
+
+> The README is very good! It gives a helpful overview of everything in the package and is clearly written, with good examples. All of the code examples worked on my computer. You might consider breaking the "experimental" parts of the package into a separate README like you do with the vignettes.
+
+- 
+
+> Link to the vignettes from the README.
+
+> The title of the github project is "A package to assist in visually testing a dataset", which doesn't describe the project that well. I suggest changing it to be the same title as on the DESCRIPTION file, which is effective in quickly summarizing what the package does.
+
+- 
+
+> You don't introduce the airquality data set at first usage, but you do introduce it after the second plot. Move the introduction to first use.
+
+> I don't understand this sentence: "When there is <0.1% of missingness, vis_miss indicates that there is >1% missingness."
+
+- Agreed, this 
+
+2. Vignettes:
+
+> addressing comments that relate to the vignette and README
+
+> Make sure you have permission to reproduce the image from R4DS. The first few paragraphs of the "Using visdat" vignette are a bit unnecessary in my opinion. You could start much more simply, with something like: "When you get a new data set, you need to get a sense of what it contains and potential problems with it." and then continue with the discussion of different ways to approach this challenge (e.g. head, glimpse, ...).
+
+- I agree with you that this is a clearer way to address this comment, and have addressed this in this commit ...
+
+> In the "Using visdat" vignette, it says "missing data represented by black", but it shows up as gray on my computer.
+
+- Interesting, it should be showing up as black, as this is what the [code and README reflect]() (well, grey50). Perhaps th
+
+> I don't know why, but in the plotly section, a vis_dat() (non-interactive) plot of df_test appears between the first two interactive plots. I can't explain it, hopefully it's just a weird quirk on my computer.
+
+- I can't seem to replicate this
+
+> There is a "future work and experimental features" section in the "Using visdat" vignette -- I suggest transferring that to the experimental vignette and just linking to it from the "Using visdat" vignette. Perhaps move the vis_guess() and the interactive plots to the experimental features vignette as well, since they seem to be evolving.
+
+- 
+
+> The plot in the experimental features vignette shows up small, I think you need to add knitr::opts_chunk$set(fig.width = 5, fig.height = 4) to the first chunk like you have in the other vignette.
+
+- Agreed, this has been completed in this commit ...
+
+# Function documentation
+
+> Did you mean to export guess_type()?
+
+- Yes, as I thought that it might be a useful function for users, but I can appreciate that perhaps it shoudl stay internal
+
+> Formatting: links, code, and other formatting need to be done with .Rd syntax. For example, for code, use \code{}, not backticks. For bold, use \strong{} instead of asterisks. See guidelines here.
+
+> Return values: The documentation doesn't include the "Value" section which is typically used to say what a function returns. I would add this. E.g. @return A \code{ggplot} object displaying the type of values in the data frame and the position of any missing values.
+
+> ?vis_compare: The documentation for this function needs to be updated. It is more a list of ideas for how to implement the function than a description of what it does.
+
+> ?vis_guess: First sentence of description seems to have an extra word or phrase, not sure exactly what was intended.
+
[…]
