### 2017-01-30 #87 @noamross
Certainly, @njtierney! That's well within expected response time.
