### 2017-02-05 579c21d
add paper response for rOpenSci review
