+++ paper/response.Rmd
@@ -15,7 +15,7 @@ And
 ...
 
 
-Consider the above my response to the positive feedback in this review. I will now respond to suggsted changes and other constructive criticisms.
+I will now respond to suggsted changes and other constructive criticisms.
 
 # @seaaan's Review
 
