### 2017-02-05 njtierney/visdat#39 @njtierney
Changes to be made in README
These are the changes to be made to the README as described in [rOpenSci onboarding](https://github.com/ropensci/onboarding/issues/87).

- [x] Introduce airquality at first usage, not after second plot (https://github.com/njtierney/visdat/commit/1985f6d1d01d3e801a098790ee23e55efd7a76c4)
- [x] Fix wording around <0.1% missingness: "When there is <0.1% of missingness, vis_miss indicates that there is >1% missingness." (https://github.com/njtierney/visdat/commit/1985f6d1d01d3e801a098790ee23e55efd7a76c4)
- [x] Link to the vignettes from the README. (https://github.com/njtierney/visdat/commit/1985f6d1d01d3e801a098790ee23e55efd7a76c4) (https://github.com/njtierney/visdat/commit/822cc87ef2249cca3be9ae8d760139ef66a645b3)
- [x] Link "experimental" parts of visdat README to "experimental" vignette (https://github.com/njtierney/visdat/commit/1985f6d1d01d3e801a098790ee23e55efd7a76c4)
- [x] Demonstrate differences between `vis_dat` and `vis_guess`, by putting them side by side by adding chunk options: `fig.show='hold', out.width='50%'`. (https://github.com/njtierney/visdat/commit/1985f6d1d01d3e801a098790ee23e55efd7a76c4)
- [x] Elaborate on what `vis_dat_ly` is doing. (https://github.com/njtierney/visdat/commit/385b52a104375bf70cb516a7c71d459d6ce13a6a)
- [x] Construct a small example of what might happen when visualising expectations. (https://github.com/njtierney/visdat/commit/445b974538099638aa7831074d99591b5de3db30)
- [x] Add guidelines for contribution in README, and CONTRIBUTION.md (https://github.com/njtierney/visdat/commit/c7ea5bb6bb12a43719fc4dbdfbccc22153072b62)
- [x] Make sure that the excessive warnings messages are fixed
