### 2017-02-05 1985f6d
README fixes relating to #39
Wording changes, links to vignette, figure size changes
