+++ vignettes/experimental_features.Rmd
@@ -9,9 +9,17 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-Shipped with visdat are some experimental features that are still a bit temperamental - `vis_compare()`, `vis_dat_ly()`, and `vis_miss_ly`. This vignette describes these features in a bit more detail.
-  
-## using `vis_compare()`
+
+Shipped with visdat are some experimental features that are still a bit temperamental:
+
+- `vis_compare()`
+- `vis_guess`,
+- `vis_dat_ly()`,
+- `vis_miss_ly`.
+
+This vignette describes these features in more detail, along with some of the plans for future work.
+
+## Using `vis_compare()`
 
 Sometimes you want to see what has changed in your data. `vis_compare()` helps with that. It is currently only just barely working, so keep in mind that this is very much in its beta stages.
 
@@ -42,4 +50,87 @@ vis_compare(iris, iris_diff_2)
 In the future I'm hoping to create some way of representing the added columns and rows as an additional facet-like part of the plot, perhaps using something from [gridExtra](https://cran.r-project.org/web/packages/gridExtra/vignettes/arrangeGrob.html).
 
 
+## Using `vis_guess()`
 
+`vis_guess()` takes a guess at what each cell is. It's best illustrated using some messy data, which we'll make here.
+
+```{r create-messy-vec}
+
+messy_vector <- c(TRUE,
+                  T,
+                  "TRUE",
+                  "T",
+                  "01/01/01",
+                  "01/01/2001",
+                  NA,
+                  NaN,
+                  "NA",
+                  "Na",
+                  "na",
+                  "10",
+                  10,
+                  "10.1",
+                  10.1,
+                  "abc",
+                  "$%TG")
+
+set.seed(1114)
+messy_df <- data.frame(var1 = messy_vector,
+                       var2 = sample(messy_vector),
+                       var3 = sample(messy_vector))
+
+```
+
+
+```{r vis-guess-messy-df, fig.show='hold', out.width='50%'}
+
+vis_guess(messy_df)
+vis_dat(messy_df)
+
+```
+
+So here we see that there are many different kinds of data in your dataframe. As an analyst this might be a depressing finding. We can see this comparison above.
+
+Here, you might just assume your data is weird because it's all factors - or worse, not notice that this is a problem.
+
+At the moment `vis_guess` is very slow. Please take this into consideration when you are using it on data with more than 1000 rows. We're looking into ways of making it faster, potentially using methods from the `parallel` package, or extending the c++ code from `readr:::collectorGuess`.
+
+# Interactivity
+
+Thanks to Carson Sievert, you can now add some really nifty interactivity into `visdat` by using `plotly::ggplotly`, allowing for information to be revealed upon mouseover of a cell. The code to do this can be seen below, but is not shown as the github README doesn't support HTML interactive graphics...yet.
+
+```{r eval = FALSE}
+
+library(plotly)
+
+vis_dat(airquality) %>% ggplotly()
+
+```
+
+## vis_dat_ly
+
+This is still under development, but it is basically a faster version of calling `vis_dat()` and then `ggplotly()`.
+
+## vis_miss_ly
+
+This is also under development, and still needs some more work on the legend.
+
+```{r vis-miss-ly}
+
+vis_miss_ly(airquality)
+
+```
+
+# Road Map
+
+**Visualising expectations**
+
+The idea here is to pass expectations into `vis_dat` or `vis_miss`, along the lines of the `expectation` command in `assertr`. For example, you could ask `vis_dat` to identify those cells with values of -1 with something like this:
+
+```{r eval = FALSE}
+
+data %>% 
+  expect(value == -1) %>%
+  vis_dat
+
+```
