### 2017-02-05 385b52a
Elaborate on `vis_dat_ly` + `expectations` to address #39
some progress made towards how the `expectations` would work
