### 2017-02-05 #87 @noamross
No specific date, but keep us appraised and if you could pick a target it
would be helpful so reviewers can plan on when to take a look at changes.

On Sun, Feb 5, 2017, 7:35 AM Nicholas Tierney <notifications@github.com>
wrote:

> Hi Everyone,
>
> I've been reading over the comments this weekend and have split the
> changes needed to be made into chunks for each section. I'm going to be
> posting these as issues and then writing commits to close the issues, so
> that I'll be able to reference exactly where/when each change from the
> reviewer comments has been made.
>
> My main response to the reviewers will be made in a single dialogue, as in
> other rOpenSci onboardings, but the issue trackers are more for my own
> benefit so that I make sure I finish everything and identify where/when I
> did so.
>
> I have not finished addressing the reviewer comments as yet, but progress
> has been made, I will update you all with the issues as I can this week.
>
> @noamross <https://github.com/noamross> is there an expected due date I
> need to finish my response by?
>
