### 2017-02-05 445b974
Demonstrate visualising expectations, addressing #39
