### 2017-02-05 c7ea5bb
Add CoC - addressing #39
