### 2017-02-12 njtierney/visdat#40 @njtierney
Changes to be made in the vignette
# Vignettes

Changes to the vignettes to address reviewer comments in [onboarding review](https://github.com/ropensci/onboarding/issues/87)

- [X] Link to vignettes (https://github.com/njtierney/visdat/commit/435154685d4f1e78dd8a03d00ba2f1b90dae39de)
- [X] introduce airquality at first usage (https://github.com/njtierney/visdat/commit/58529ef005297b1fe5a3e412e4a1f983f631234d)
- [X] sentence about <0.1% of missingness and >1% missingness (https://github.com/njtierney/visdat/commit/58529ef005297b1fe5a3e412e4a1f983f631234d)
- [X] Change figure size in experimental vignette: `knitr::opts_chunk$set(fig.width = 5, fig.height = 4)` (https://github.com/njtierney/visdat/commit/bcb01a760d36ad821e7fd9eaa4463ffa3daa3b2c)
- [X] move `visdat_ly()` and `vis_miss_ly()` to the experimental vignette. (https://github.com/njtierney/visdat/commit/435154685d4f1e78dd8a03d00ba2f1b90dae39de)
- [x] Make sure that the excessive warning messages are fixed
- [x] Decide whether I really need to discuss the figure from R4DS.
