### 2017-02-12 4351546
link to vignettes from README, addressing #39
