+++ vignettes/experimental_features.Rmd
@@ -97,16 +97,6 @@ At the moment `vis_guess` is very slow. Please take this into consideration when
 
 # Interactivity
 
-Thanks to Carson Sievert, you can now add some really nifty interactivity into `visdat` by using `plotly::ggplotly`, allowing for information to be revealed upon mouseover of a cell. The code to do this can be seen below, but is not shown as the github README doesn't support HTML interactive graphics...yet.
-
-```{r eval = FALSE}
-
-library(plotly)
-
-vis_dat(airquality) %>% ggplotly()
-
-```
-
 ## vis_dat_ly
 
 `vis_dat_ly(airquality)` is a faster version of calling `ggplotly(vis_dat(airquality))`. When hovering the cursor over a cell, it will display information about each each cell, the type of data, the value, the row number, and the variable.
