+++ vignettes/using_visdat.Rmd
@@ -169,86 +169,7 @@ vis_miss(airquality) %>% ggplotly()
 
 ```
 
-There is also an in-development feature, `vis_miss_ly`, which directly uses plotly's code, which is faster than calling ggplotly(vis_dat(df)).
+# Future work
 
-```{r}
-
-vis_miss_ly(airquality)
-
-```
-
-Although some work is needed still to fix the legend and fill, and add percentages to the legend.
-
-## using `vis_guess`
-
-`vis_guess` takes a guess at what each cell is. It's best illustrated using some messy data, which we'll make here.
-
-```{r create-messy-vector}
-
-messy_vector <- c(TRUE,
-                  T,
-                  "TRUE",
-                  "T",
-                  "01/01/01",
-                  "01/01/2001",
-                  NA,
-                  NaN,
-                  "NA",
-                  "Na",
-                  "na",
-                  "10",
-                  10,
-                  "10.1",
-                  10.1,
-                  "abc",
-                  "$%TG")
-
-messy_df <- data.frame(var1 = messy_vector,
-                       var2 = sample(messy_vector),
-                       var3 = sample(messy_vector))
-
-```
-
-Here if we just glimpse `messy_df`, we get 
-
-```{r}
-
-dplyr::glimpse(messy_df)
-
-```
-
-It looks like we have 3 factors. But this seems a bit strange. `vis_guess` takes a guess at what each element is, visualising it for you:
-
-```{r messy-vis-guess}
-
-vis_guess(messy_df)
-
-```
-
-So here we see that there are many different kinds of data in your dataframe. As an analyst this might be a depressing finding. Compare this to `vis_dat`.
-
-```{r visdat-messy}
-
-vis_dat(messy_df)
-
-```
-
-Where you'd just assume your data is weird because it's all factors - or worse, not notice that this is a problem.
-
-At the moment `vis_guess` is slow. Please take this into consideration when you are using it on data with more than 1000 rows. We're looking into ways of making it faster, potentially using methods from the `parallel` package, or extending the c++ code from `readr::parse_guess`.
-
-
-# Future work and experimental features
-
-**visualising expectations**
-
-The idea is to pass expectations into `vis_dat` or `vis_miss`, along the lines of the `expectation` command in `assertr`. For example, you could ask `vis_dat` to identify those cells with values of -1 with something like this:
-
-```{r eval = FALSE}
-
-data %>% 
-  expect(value == -1) %>%
-  vis_dat
-
-```
+There are some other features in development for interactivity, guessing types of data, and comparing datasets. You read more about these in the vignette, ["Experimental Features of visdat"](http://www.njtierney.com/visdat/articles/experimental_features.html).
 
