### 2017-02-12 bbd098d
better describe the airquality dataset, #39, #40
