### 2017-02-12 bcb01a7
fix vignette figure size, #40
