### 2017-02-12 njtierney/visdat#41 @njtierney
Changes to documentation
# Documentation

Address suggested changes to documentation, as kindly provided in https://github.com/ropensci/onboarding/issues/87

- [X] correct use of roxygen links, code, emphasis, and seeAlso (`\code{\link[visdat]{vis_miss}}`)  (https://github.com/njtierney/visdat/commit/05ab39273408bad8c6d3c5ff447f1da7849694ec)
- [X] add return return value in roxygen, e.g., `@return A \code{ggplot} object displaying the type of values in the data frame and the position of any missing values.` (https://github.com/njtierney/visdat/commit/05ab39273408bad8c6d3c5ff447f1da7849694ec)
- [X] Update vis_compare documentation, change from list of ideas to description (https://github.com/njtierney/visdat/commit/dbdbb7e39fe9013108382aa0bff99dbe13dac4ce)
- [X] vis_guess, fix first sentence (https://github.com/njtierney/visdat/commit/b1e33d7b213f677d606ab14fcb641eec49dc0b0c)
- [X] Add content to roxygen file for `visdat` so that `?visdat` provides documentation (https://github.com/njtierney/visdat/commit/fb29a39c90934a3507b43e9450328bf459e0b3f7)
