### 2017-02-13 #87 @njtierney
Hi Everyone, just wanted to give y'all an update. Basically, I'm just allocating tasks to myself in issues, then describing the changes with reference to the commits and issues in my response to the reviews. I figure this way people can see exactly what has happened where.

I'm in the middle of responding to comments on two thesis papers, but I'm going to make an effort to knock over a little piece of visdat every day this week. By Friday I should have a better idea of where this is at.
