+++ paper/response.Rmd
@@ -12,8 +12,6 @@ And
 
 >
 
-...
-
 
 I will now respond to suggsted changes and other constructive criticisms.
 
@@ -23,19 +21,23 @@ I will now respond to suggsted changes and other constructive criticisms.
 
 > The README is very good! It gives a helpful overview of everything in the package and is clearly written, with good examples. All of the code examples worked on my computer. You might consider breaking the "experimental" parts of the package into a separate README like you do with the vignettes.
 
-- 
+- Agreed, but I would prefer to keep the README as a whole piece, so I think instead I will just link to the experimental vignette as in the README. This has been logged in [issue 39]() and fixed in commit 
 
 > Link to the vignettes from the README.
 
+- Agreed, this has been address in commit ..., issue ...
+
 > The title of the github project is "A package to assist in visually testing a dataset", which doesn't describe the project that well. I suggest changing it to be the same title as on the DESCRIPTION file, which is effective in quickly summarizing what the package does.
 
-- 
+- Agreed, this makes more sense - changed!
 
 > You don't introduce the airquality data set at first usage, but you do introduce it after the second plot. Move the introduction to first use.
 
+Fixed in ...
+
 > I don't understand this sentence: "When there is <0.1% of missingness, vis_miss indicates that there is >1% missingness."
 
-- Agreed, this 
+- Agreed, this has been changed in issue ....
 
 2. Vignettes:
 
@@ -47,27 +49,29 @@ I will now respond to suggsted changes and other constructive criticisms.
 
 > In the "Using visdat" vignette, it says "missing data represented by black", but it shows up as gray on my computer.
 
-- Interesting, it should be showing up as black, as this is what the [code and README reflect]() (well, grey50). Perhaps th
+- Interesting, it should be showing up as black, as this is what the [code and README reflect]() (well, grey50). Does it still happen now?
 
 > I don't know why, but in the plotly section, a vis_dat() (non-interactive) plot of df_test appears between the first two interactive plots. I can't explain it, hopefully it's just a weird quirk on my computer.
 
-- I can't seem to replicate this
+- I can't seem to replicate this, can you maybe take a screenshot of this?
 
 > There is a "future work and experimental features" section in the "Using visdat" vignette -- I suggest transferring that to the experimental vignette and just linking to it from the "Using visdat" vignette. Perhaps move the vis_guess() and the interactive plots to the experimental features vignette as well, since they seem to be evolving.
 
-- 
+- Agreed, this has been done in commit ... , issue ...
 
 > The plot in the experimental features vignette shows up small, I think you need to add knitr::opts_chunk$set(fig.width = 5, fig.height = 4) to the first chunk like you have in the other vignette.
 
-- Agreed, this has been completed in this commit ...
+- Agreed, this has been completed in this commit ..., issue...
 
 # Function documentation
 
 > Did you mean to export guess_type()?
 
-- Yes, as I thought that it might be a useful function for users, but I can appreciate that perhaps it shoudl stay internal
+- Yes, as I thought that it might be a useful function for users, although I can't think of a good usecase outside of `vis_guess` right now. Perhaps it might be best for `guess_type` to be unexported?
 
 > Formatting: links, code, and other formatting need to be done with .Rd syntax. For example, for code, use \code{}, not backticks. For bold, use \strong{} instead of asterisks. See guidelines here.
+
+
 
 > Return values: The documentation doesn't include the "Value" section which is typically used to say what a function returns. I would add this. E.g. @return A \code{ggplot} object displaying the type of values in the data frame and the position of any missing values.
 
@@ -150,11 +154,15 @@ I will now respond to suggsted changes and other constructive criticisms.
 
[…]
