### 2017-02-15 9e909e1
change to markdown roxygen syntax
