+++ R/miss_guide_label.R
@@ -1,10 +1,10 @@
 #' Label the legend with the percent of missing data
 #'
-#' \code{miss_guide_label} is an internal function for vis_miss to label the legend.
+#' `miss_guide_label` is an internal function for vis_miss to label the legend.
 #'
 #' @param x is a dataframe passed from vis_miss(x).
 #'
-#' @return a data_frame with two columns \code{p_miss_lab} and \code{p_pres_lab}, containing the labels to use for present and missing. A dataframe is returned because I think it is a good style habit compared to a list.
+#' @return a data_frame with two columns `p_miss_lab` and `p_pres_lab`, containing the labels to use for present and missing. A dataframe is returned because I think it is a good style habit compared to a list.
 #'
 #'
 miss_guide_label <- function(x) {
