+++ R/vis_guess.R
@@ -1,6 +1,6 @@
 #' Visualise a data.frame like vis_dat, but takes a guess at to telling you what it contains.
 #'
-#' \code{vis_guess} just like vis_dat, vis_guess gives you an at-a-glance ggplot object of what is inside a dataframe. Except this time, vis_guess uses `readr` to help it guess what the individual cells are. These cells are coloured according to what class they are and whether the values are missing. As \code{vis_guess} returns a ggplot object, it is very easy to customize. Currently very slow.
+#' `vis_guess` just like vis_dat, vis_guess gives you an at-a-glance ggplot object of what is inside a dataframe. Except this time, vis_guess uses `readr` to help it guess what the individual cells are. These cells are coloured according to what class they are and whether the values are missing. As `vis_guess` returns a ggplot object, it is very easy to customize. Currently very slow.
 #'
 #' @param x a data.frame object
 #'
