+++ DESCRIPTION
@@ -8,7 +8,7 @@ Depends:
     R (>= 3.2.2)
 License: MIT + file LICENSE
 LazyData: true
-RoxygenNote: 6.0.0
+RoxygenNote: 6.0.1.9000
 Imports:
     ggplot2,
     tidyr,
@@ -26,3 +26,4 @@ Suggests:
     rmarkdown,
     vdiffr
 VignetteBuilder: knitr
+Roxygen: list(markdown = TRUE)

[…]
