### 2017-02-15 fb29a39
add documentation for ?visdat
