### 2017-02-15 dbdbb7e
fix vis_compare documentation
