### 2017-02-15 b1e33d7
make vis_guess documentation clearer
