### 2017-02-15 njtierney/visdat#41 @njtierney
Closed by https://github.com/njtierney/visdat/commit/7e4c74a2531e5ab5cd5d57feca1f3372271846a3
