### 2017-02-15 njtierney/visdat#43 @njtierney
Changes to be made in Examples
# Examples

As kindly described in the rOpenSci onboarding: addressing review in https://github.com/ropensci/onboarding/issues/87

- [x] Don't need to call `library(visdat)` in examples (https://github.com/njtierney/visdat/commit/1afb83e1114af93dd31054da3a4cbc4fda39e9ef)
- [x] palette is misspelled (https://github.com/njtierney/visdat/commit/31909e241c1e1cc1b640c55edaea9c228f92976e)
- [x] fix warning messages in `vis_compare` (https://github.com/njtierney/visdat/commit/6564ea25067f272dbd7d31cfacb6cc3a95081767)
- [x] fix warning messages in `vis_guess` (https://github.com/njtierney/visdat/commit/6564ea25067f272dbd7d31cfacb6cc3a95081767)
- [x] replace `purrr::dmap` with `purrr::map_df` (https://github.com/njtierney/visdat/commit/6564ea25067f272dbd7d31cfacb6cc3a95081767)
- [x] does `vis_compare` cluster by default?

From what I can see, vis_compare does not cluster by default, that was just due to the example that I gave.

``` r
library(visdat)
# make a new dataset of iris that contains some NA values
iris_diff <- iris
iris_diff[sample(x = 1:nrow(iris), 10), sample(1:4,2)] <- NA

vis_compare(iris, iris_diff)
#> vis_compare is in BETA! If you have suggestions or errors
#> post an issue at https://github.com/njtierney/visdat/issues
#> Warning: attributes are not identical across measure variables; they will
#> be dropped

#> Warning: attributes are not identical across measure variables; they will
#> be dropped
```

![](http://i.imgur.com/YPzJByp.png)

Ooops, this reminds me I need to fix suppress those warnings
