### 2017-02-16 1afb83e
remove calls to library(visdat) - addressing #43
