+++ R/vis_dat.R
@@ -16,8 +16,6 @@
 #'
 #' @examples
 #'
-#' library(visdat)
-#'
 #' vis_dat(airquality)
 #'
 #' # flip the axis to look more like a dataframe
