+++ R/vis_miss.R
@@ -18,8 +18,6 @@
 #'
 #' @examples
 #'
-#' library(visdat)
-#'
 #' vis_miss(airquality)
 #'
 #' vis_miss(airquality, cluster = TRUE)
