### 2017-02-16 31909e2
fix spelling of palette
