+++ R/vis_dat.R
@@ -56,7 +56,7 @@ vis_dat <- function(x,
   # reshape the dataframe ready for geom_raster
   d <- x %>%
     # mutate_each_(funs(fingerprint), tbl_vars(.)) %>%
-    purrr::dmap(fingerprint) %>%
+    purrr::map_df(fingerprint) %>%
     dplyr::mutate(rows = seq_len(nrow(.))) %>%
     tidyr::gather_(key_col = "variables",
                    value_col = "valueType",
