### 2017-02-17 njtierney/visdat#44 @njtierney
Address changes in functionality
These were kindly provided in the rOpenSci onboarding review - https://github.com/ropensci/onboarding/issues/87

- [x] change `purrr::dmap` to `purrr::map_df` to remove warnings (https://github.com/njtierney/visdat/commit/6564ea25067f272dbd7d31cfacb6cc3a95081767)

- [X]  suppress the warning message and add helper function for: (https://github.com/njtierney/visdat/commit/8cd0bfd33699dd6e0bf1de2e45b82b091f2fb1bf)

```
  d$value <- tidyr::gather_(x, "variables", "value", names(x))$value
```

- [X] suppress warning message: `("Warning message: attributes are not identical across measure variables; they will be dropped")` from the call `tidyr::gather_(x, "variables", "value", names(x))$value`. This message is not relevant to the end user and should be suppressed. (https://github.com/njtierney/visdat/commit/8cd0bfd33699dd6e0bf1de2e45b82b091f2fb1bf)
- [x] duplication in code: (https://github.com/njtierney/visdat/commit/0a6b459873a3183f83ee4c290144b625a2a0f66c)

```
dplyr::mutate(rows = seq_len(nrow(.))) %>%
  tidyr::gather_(key_col = "variables",
                 value_col = "valueType",
                 gather_cols = names(.)[-length(.)])
```

- [x] remove duplication of flipping rows:

```
 if (flip == TRUE){
    suppressMessages({
      vis_dat_plot +
        ggplot2::scale_y_reverse() +
        ggplot2::scale_x_discrete(position = "top",
                                  limits = type_order_index) +
        ggplot2::theme(axis.text.x = ggplot2::element_text(hjust = 0.5)) +
        ggplot2::labs(x = "")
    })
  } else if (flip == FALSE){
    vis_dat_plot
  }

```

- [x] Add helper function for the plots, since they seem to be the same or very similar. Perhaps this could be named something like: `create_vis_`
- [x] Put the rows in `vis_` and friends as they would be in a dataframe, so it is more intuitive. Note: This was the original use of `flip`, so that by default the visualisations "read" from top to bottom. This should extend to all plots.
- [ ] Should the rows and columns appear to be in same order as the input - `sort_type` should be off by default.
- [x] Should the titles at the bottom of the columns make sense?
- [x] turn `if (sort_type == TRUE)` into `if (sort_type)`, and `if (x == FALSE)` is into `if (!x)`.
- [x] Move `vis_compare` and `vis_guess` should to a development version?
