### 2017-02-17 0a6b459
add vis_gather_ to remove duplication in the gathering step of vis_* fam
