+++ R/internals.R
@@ -1,0 +1,13 @@
+#' (Internal) Gather rows into a format appropriate for grid visualisation
+#'
+#' @param x a dataframe
+#'
+#' @return data.frame gathered to have columns "variables", "valueType", and a row id called "rows".
+#'
+vis_gather_ <- function(x){
+  x %>%
+  dplyr::mutate(rows = seq_len(nrow(x))) %>%
+    tidyr::gather_(key_col = "variables",
+                   value_col = "valueType",
+                   gather_cols = names(.)[-length(.)])
+}
