+++ man/vis_gather_.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/internals.R
+\name{vis_gather_}
+\alias{vis_gather_}
+\title{(Internal) Gather rows into a format appropriate for grid visualisation}
+\usage{
+vis_gather_(x)
+}
+\arguments{
+\item{x}{a dataframe}
+}
+\value{
+data.frame gathered to have columns "variables", "valueType", and a row id called "rows".
+}
+\description{
+(Internal) Gather rows into a format appropriate for grid visualisation
+}
