### 2017-02-17 93bbf96
add test for vis_gather to ensure it returns same output as old method
