### 2017-02-17 20c1a2b
fix strange pop stash - ensure vis_gather_ is called properly
