### 2017-02-17 8cd0bfd
add internal fun to add values from gathered data & suppress warnings
