+++ R/internals.R
@@ -11,3 +11,23 @@ vis_gather_ <- function(x){
                    value_col = "valueType",
                    gather_cols = names(.)[-length(.)])
 }
+
+
+#' (Internal) Add values of each row as a column
+#'
+#' This adds information about each row, so that when called by plotly, the values are made visible on hover. Warnings are suppressed because tidyr gives a warning about type coercion, which is fine.
+#'
+#' @param x dataframe created from `vis_gather_`
+#'
+#' @return the x dataframe with the added column `value`.
+#'
+vis_extract_value_ <- function(x){
+
+  suppressWarnings(
+    tidyr::gather_(x,
+                   "variables",
+                   "value",
+                   names(x))$value
+  )
+
+}
