+++ man/vis_extract_value_.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/internals.R
+\name{vis_extract_value_}
+\alias{vis_extract_value_}
+\title{(Internal) Add values of each row as a column}
+\usage{
+vis_extract_value_(x)
+}
+\arguments{
+\item{x}{dataframe created from \code{vis_gather_}}
+}
+\value{
+the x dataframe with the added column \code{value}.
+}
+\description{
+This adds information about each row, so that when called by plotly, the values are made visible on hover. Warnings are suppressed because tidyr gives a warning about type coercion, which is fine.
+}
