### 2017-02-17 f93be0f
add test for new function - vis_gather_value_
