### 2017-02-26 e4af28a
added boilerplate for vis_create
