+++ R/internals.R
@@ -31,3 +31,26 @@ vis_extract_value_ <- function(x){
   )
 
 }
+
+#' (Internal) Create a boilerploate for visualisations of the vis_ family
+#'
+#' @param x a dataframe in longformat as transformed by `vis_gather_` and `vis_extract_value`.
+#'
+#' @return a ggplot object
+#'
+vis_create_ <- function(x){
+
+  ggplot2::ggplot(data = x,
+                  ggplot2::aes_string(x = "variables",
+                                      y = "rows",
+                                    # text assists with plotly mouseover
+                                    text = "value")) +
+  ggplot2::geom_raster(ggplot2::aes_string(fill = "valueType")) +
+  ggplot2::theme_minimal() +
+  ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 45,
+                                                     vjust = 1,
+                                                     hjust = 1)) +
+  ggplot2::labs(x = "Variables in Dataset",
+                y = "Observations")
+
+}
