### 2017-02-26 255bfec
boilerplate doesn't quite work for vis_compare
This is due to the invlusion of the aes(text = c("value...") stuff. I wonder if I can just add this at the end though?
