### 2017-02-26 c2d5457
add the boilerplate for vis_dat, vis_guess, vis_miss
