+++ R/vis_dat.R
@@ -64,18 +64,9 @@ vis_dat <- function(x,
 
   # do the plotting
   vis_dat_plot <-
-    ggplot2::ggplot(data = d,
-                    ggplot2::aes_string(x = "variables",
-                    y = "rows",
-                    # text assists with plotly mouseover
-                    text = "value")) +
-    ggplot2::geom_raster(ggplot2::aes_string(fill = "valueType")) +
-    ggplot2::theme_minimal() +
-    ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 45,
-                                     vjust = 1,
-                                     hjust = 1)) +
-    ggplot2::labs(x = "Variables in Dataset",
-         y = "Observations") +
+    # add the boilerplate
+    vis_create_(d) +
+    # change the limits etc.
     ggplot2::scale_x_discrete(limits = type_order_index) +
     ggplot2::guides(fill = ggplot2::guide_legend(title = "Type"))
 
