### 2017-02-26 41de92f
boilerplate vis_create_ doco
