+++ man/vis_create_.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/internals.R
+\name{vis_create_}
+\alias{vis_create_}
+\title{(Internal) Create a boilerploate for visualisations of the vis_ family}
+\usage{
+vis_create_(x)
+}
+\arguments{
+\item{x}{a dataframe in longformat as transformed by \code{vis_gather_} and \code{vis_extract_value}.}
+}
+\value{
+a ggplot object
+}
+\description{
+(Internal) Create a boilerploate for visualisations of the vis_ family
+}
