### 2017-02-26 2471c46
flip axes by default - remove options to flip axes.
