### 2017-02-26 488a779
add a linea break and some space between the legend labels
