### 2017-02-26 d3bd44d
move more of the axis flipping parts into vis_create_
