+++ R/internals.R
@@ -51,6 +51,9 @@ vis_create_ <- function(x){
                                                      vjust = 1,
                                                      hjust = 1)) +
   ggplot2::labs(x = "Variables in Dataset",
-                y = "Observations")
+                y = "Observations") +
+    # flip the axes
+    ggplot2::scale_y_reverse() +
+    ggplot2::theme(axis.text.x = ggplot2::element_text(hjust = 0.5))
 
 }
