### 2017-02-26 3439039
remove == TRUE
