### 2017-02-26 4217ca1
add internal fun for creating % missing for each col
