+++ R/internals.R
@@ -57,3 +57,29 @@ vis_create_ <- function(x){
     ggplot2::theme(axis.text.x = ggplot2::element_text(hjust = 0.5))
 
 }
+
+#' (Internal) Create labels for the columns containing the \% missing data
+#'
+#' @param x data.frame
+#' @param col_order_index the order of the columns
+#'
+#' @return data.frame containing the missingness percent down to 0.1 percent
+#'
+#' @export
+label_col_missing_pct <- function(x,
+                                  col_order_index){
+
+  # present everything in the right order
+  purrr::map_df(x, ~round(mean(is.na(.))*100,1))[col_order_index] %>%
+    purrr::map_chr(function(x){
+      dplyr::case_when(
+        x == 0 ~  "0%",
+        x > 0.1 ~ paste0(x,"%"),
+        x < 0.1 ~ "<0.1%"
+      )
+    }) %>%
+    paste0(col_order_index,
+           "\n",
+           .)
+
+}
