+++ man/label_col_missing_pct.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/internals.R
+\name{label_col_missing_pct}
+\alias{label_col_missing_pct}
+\title{(Internal) Create labels for the columns containing the \% missing data}
+\usage{
+label_col_missing_pct(x, col_order_index)
+}
+\arguments{
+\item{x}{data.frame}
+
+\item{col_order_index}{the order of the columns}
+}
+\value{
+data.frame containing the missingness percent down to 0.1 percent
+}
+\description{
+(Internal) Create labels for the columns containing the \% missing data
+}
