### 2017-02-26 709c0b2
add show_perc_col arg to vis_miss
