+++ R/vis_miss.R
@@ -4,12 +4,13 @@
 #'
 #' @param x a data.frame
 #'
-#' @param cluster logical TRUE/FALSE. TRUE specifies that you want to use hierarchical clustering (mcquitty method) to arrange rows according to missingness. FALSE specifies that you want to leave it as is
+#' @param cluster logical. TRUE specifies that you want to use hierarchical clustering (mcquitty method) to arrange rows according to missingness. FALSE specifies that you want to leave it as is
 #'
-#' @param sort_miss logical TRUE/FALSE. TRUE arranges the columns in order of missingness
+#' @param sort_miss logical. TRUE arranges the columns in order of missingness
 #'
-#' @param show_perc logical TRUE/FALSE. TRUE now adds in the \% of missing/complete data in the whole dataset into the legend. Default value is TRUE
+#' @param show_perc logical. TRUE now adds in the \% of missing/complete data in the whole dataset into the legend. Default value is TRUE
 #'
+#'@param show_perc_col logical. TRUE adds in the \% missing data in a given column into the x axis. Can be disabled if
 #'
 #' @return `ggplot2` object displaying the position of missing values in the dataframe, and the percentage of values missing and present.
 #'
@@ -27,10 +28,10 @@
 vis_miss <- function(x,
                      cluster = FALSE,
                      sort_miss = FALSE,
-                     show_perc = TRUE){
+                     show_perc = TRUE,
+                     show_perc_col = TRUE){
   # make a TRUE/FALSE matrix of the data.
   # This tells us whether it is missing (true) or not (false)
-  # x = airquality
   x.na <- is.na(x)
 
   # switch for creating the missing clustering
@@ -110,16 +111,29 @@ vis_miss <- function(x,
   }
 
     # then we plot it
-  vis_create_(d) +
+  vis_miss_plot <- vis_create_(d) +
       ggplot2::scale_fill_manual(name = "",
                         values = c("grey80",
                                    "grey20"),
                         labels = c(p_pres_lab,
                                      p_miss_lab)) +
       ggplot2::guides(fill = ggplot2::guide_legend(reverse = TRUE)) +
+    ggplot2::theme(legend.position = "bottom")
+
+  # add the missingness column labels
+  if (show_perc_col){
+
     # flip the axes, add the info about limits
+    vis_miss_plot +
+      ggplot2::scale_x_discrete(position = "top",
+                              limits = col_order_index,
+                              labels = label_col_missing_pct(x,
+                                                             col_order_index))
+  } else {
+    vis_miss_plot +
     ggplot2::scale_x_discrete(position = "top",
                               limits = col_order_index)
+  }
 
       # guides(fill = guide_legend(title = "Type"))
   # Thanks to
