### 2017-02-26 8a32d75
remove extra linebreak from vis_miss legend
