+++ R/miss_guide_label.R
@@ -32,12 +32,12 @@ miss_guide_label <- function(x) {
     # create the labels
     p_miss_lab <- paste("Missing \n(",
                         p_miss,
-                        "%)\n",
+                        "%)",
                         sep = "")
 
     p_pres_lab <- paste("Present \n(",
                         p_pres,
-                        "%)\n",
+                        "%)",
                         sep = "")
     }
 
