### 2017-02-26 400cc44
vis_miss now shows the % missing in each column
