+++ NAMESPACE
@@ -1,6 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 export(guess_type)
+export(label_col_missing_pct)
 export(vis_compare)
 export(vis_dat)
 export(vis_guess)
