### 2017-03-13 #87 @njtierney
Hi Everyone,

Just wanted to touch base with another update. I'm not far off from submitting my thesis, and this is where all my energy is going at the moment - I am hopeful that I can address the remaining issues soon, but I just wanted to let you know that finishing this review is definitely on my mind, and that I do value this review process very much!

Hopefully I can have my reply to this in the next week or two, please let me know if you need it sooner and I'll see if I can juggle a few things.
