### 2017-05-10 #87 @njtierney
Hi everyone!

I'm back from outer space, and so working on visdat is now a priority!

I will aim to have this review complete by Friday, I look forward to delivering it to you all!
