### 2017-05-10 822cc87
Add both vignettes, addressing issue #39
