+++ README.Rmd
@@ -41,9 +41,11 @@ There are two main commands in the `visdat` package:
 
 - `vis_dat()` visualises a dataframe showing you what the classes of the columns are, and also displaying the missing data.
 
-- `vis_miss()` visualises just the missing data, and allows for missingness to be clustered and columns rearranged. `vis_miss()` is similar to `missing.pattern.plot` from the [`mi`](https://cran.r-project.org/web/packages/mi/index.html) package. Unfortunately `missing.pattern.plot` is no longer in the `mi` package (well, as of 14/02/2016).
+- `vis_miss()` visualises just the missing data, and allows for missingness to be clustered and columns rearranged. `vis_miss()` is similar to `missing.pattern.plot` from the [`mi`](https://cran.r-project.org/web/packages/mi/index.html) package. Unfortunately `missing.pattern.plot` is no longer in the `mi` package (as of 14/02/2016).
 
-There are also experimental which are described in more detail in the [vignette: "Experimental Features of visdat](http://www.njtierney.com/visdat/articles/experimental_features.html).
+You can read more about visdat in the vignette, ["using visdat""](http://www.njtierney.com/visdat/articles/using_visdat.html).
+
+There are also experimental which are described in more detail in the vignette ["Experimental Features of visdat](http://www.njtierney.com/visdat/articles/experimental_features.html).
 
 Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
 
@@ -72,7 +74,6 @@ vis_dat(airquality,
         sort_type = FALSE)
 
 ```
-
 
 ### With many kinds of data
 
