+++ man/visdat.Rd
@@ -6,7 +6,10 @@
 \alias{visdat-package}
 \title{visdat}
 \description{
-visdat is a package that helps with the preliminary visualisation of data. visdat makes it easy to visualise your whole dataset so that you can visually identify problems. It's main functions are:
+visdat is a package that helps with the preliminary visualisation of data. visdat makes it easy to visualise your whole dataset so that you can visually identify problems.
+}
+\seealso{
+It's main functions are:
 \itemize{
 \item \code{\link[=vis_dat]{vis_dat()}}
 \item \code{\link[=vis_miss]{vis_miss()}}
@@ -14,4 +17,6 @@ visdat is a package that helps with the preliminary visualisation of data. visda
 \item \code{\link[=vis_compare]{vis_compare()}}
 \item \code{\link[=vis_miss_ly]{vis_miss_ly()}}
 }
+
+To learn more about visdat, see the vignette at \url{www.njtierney.com/visdat/articles/using_visdat.html}
 }
