### 2017-05-10 3f74141
all chunks have names, remove ref/pic in r4ds
