### 2017-05-10 d71d2ac
removed library call in examples
