### 2017-05-10 0b837d7
update paper and response
