+++ paper/paper.Rmd
@@ -20,13 +20,7 @@ output:
 
 # Summary
 
-As described by Hadley Wickham [@wickham2016], there are 6 phases of data science:
-
-![Process of Data Science](http://r4ds.had.co.nz/diagrams/data-science.png)
-
-You can get insight into your data by modelling, visualising, and transforming, which Wickham describes as "Understanding" or "knowledge generation". There is some overlap here, as in this process it wouldn't be surprising that you might uncover some feature of your dataset that you would need to clean up or change. For example, in analysing some data, you might discover that some strings were indeed factors, or that gender was considered numeric in your regression, warranting you to re-visit the Tidying phase.
-
-In the same way, this process gets applied in the Tidying step of data science. You read in your data, but you then need to look at it to understand what you need to do to make it ready for analysis. A phrase that often arises in this process is "looking at the data". It is helpful to look more closely at what this means. You can look at the first six rows of data, the head of the data:
+When you get a new data set, you need to look at the data to get a sense of what it contains and potential problems with it. You read in your data, but you then need to look at it to understand what you need to do to make it ready for analysis. A phrase that often arises in this process is "looking at the data". It is helpful to look more closely at what this means. You can look at the first six rows of data, the head of the data:
 
 ```{r head-iris}
 
