### 2017-05-12 fb191d8
various updates to the rOpenSci response
