+++ paper/response.Rmd
@@ -2,15 +2,15 @@
 
 I would like to thank the reviewers for taking the time to review visdat so thoroughly. It is fantastic to hear such encouraging comments from members of the community.
 
-I will now respond to suggested changes and other constructive criticisms.
+Below I respond to suggested changes and other constructive criticisms.
 
 # @seaaan's Review
 
 1. Documentation: README
 
-> The README is very good! It gives a helpful overview of everything in the package and is clearly written, with good examples. All of the code examples worked on my computer. You might consider breaking the "experimental" parts of the package into a separate README like you do with the vignettes.
+> You might consider breaking the "experimental" parts of the package into a separate README like you do with the vignettes.
 
-- Agreed, but I would prefer to keep the README as a whole piece, so I think instead I will just link to the experimental vignette as in the README. This has been logged in [issue 39](https://github.com/njtierney/visdat/issues/39) and fixed in the indicated commit messages.
+- Agreed! I would prefer to keep the README as a whole piece, Perhaps instead I could just link to the experimental vignette as in the README. This has been logged in [issue 39](https://github.com/njtierney/visdat/issues/39) and fixed in the indicated commit messages.
 
 > Link to the vignettes from the README.
 
@@ -32,7 +32,7 @@ I will now respond to suggested changes and other constructive criticisms.
 
 > Make sure you have permission to reproduce the image from R4DS. The first few paragraphs of the "Using visdat" vignette are a bit unnecessary in my opinion. You could start much more simply, with something like: "When you get a new data set, you need to get a sense of what it contains and potential problems with it." and then continue with the discussion of different ways to approach this challenge (e.g. head, glimpse, ...).
 
-- I agree with you that this is a clearer way to address this comment, and have addressed this in this commit ...
+- I agree with you that this is a clearer way to introduce the topic - this is now [changed](https://github.com/njtierney/visdat/commit/3f74141d2e9de4bc0ae870fc874b17a21829618e).
 
 > In the "Using visdat" vignette, it says "missing data represented by black", but it shows up as gray on my computer.
 
@@ -40,15 +40,15 @@ I will now respond to suggested changes and other constructive criticisms.
 
 > I don't know why, but in the plotly section, a vis_dat() (non-interactive) plot of df_test appears between the first two interactive plots. I can't explain it, hopefully it's just a weird quirk on my computer.
 
-- I can't seem to replicate this, can you maybe take a screenshot of this?
+- I can't seem to replicate this, can you maybe take a screenshot of this if this is still happening?
 
 > There is a "future work and experimental features" section in the "Using visdat" vignette -- I suggest transferring that to the experimental vignette and just linking to it from the "Using visdat" vignette. Perhaps move the vis_guess() and the interactive plots to the experimental features vignette as well, since they seem to be evolving.
 
-- Agreed, I have now linked to this additional experimental vignette in the "using visdat" vignette.
+- Great point! Thank you! I have now linked to this additional experimental vignette in the "using visdat" vignette.
 
 > The plot in the experimental features vignette shows up small, I think you need to add knitr::opts_chunk$set(fig.width = 5, fig.height = 4) to the first chunk like you have in the other vignette.
 
-- Agreed, this has been addressed
+- Good idea! This has been addressed now.
 
 # Function documentation
 
@@ -58,7 +58,7 @@ I will now respond to suggested changes and other constructive criticisms.
 
 > Formatting: links, code, and other formatting need to be done with .Rd syntax. For example, for code, use \code{}, not backticks. For bold, use \strong{} instead of asterisks.
 
-- I have now converted this over to use markdown with roxygen.
+- I have now converted this over to use markdown with roxygen, this should be fixed now.
[…]
