### 2017-05-12 fa65a44
vis_* family now has the labels on the top of the plot
