### 2017-06-08 e8aa445
Update the axis printing to avoid clutter
