+++ R/vis_miss.R
@@ -126,9 +126,11 @@ vis_miss <- function(x,
     # flip the axes, add the info about limits
     vis_miss_plot +
       ggplot2::scale_x_discrete(position = "top",
-                              limits = col_order_index,
-                              labels = label_col_missing_pct(x,
-                                                             col_order_index))
+                                limits = col_order_index,
+                                labels = label_col_missing_pct(x,
+                                                               col_order_index)) +
+      # fix up the location of the text
+      ggplot2::theme(axis.text.x = ggplot2::element_text(hjust = 0.2))
   } else {
     vis_miss_plot +
     ggplot2::scale_x_discrete(position = "top",
