### 2017-06-08 93ae53f
remove the flip argument
