+++ vignettes/using_visdat.Rmd
@@ -157,7 +157,7 @@ Thanks to Carson Sievert, you can now add some really nifty interactivity into v
 
 library(plotly)
 
-vis_dat(airquality, flip = TRUE) %>% ggplotly()
+vis_dat(airquality) %>% ggplotly()
 vis_miss(airquality) %>% ggplotly()
 
 ```
