+++ NEWS.md
@@ -4,6 +4,7 @@
 
 - axes in `vis_` family are now flipped by default
 - `vis_miss` now shows the % missingness in a column, can be disabled by setting `show_perc_col` argument to FALSE
+- removed `flip` argument, as this should be the default 
 
 ## Minor Improvements
 
