### 2017-06-08 48a3a4f
update the alignment so that it is directly above the column of interest
