### 2017-06-17 1ded5e7
rOpenSci checks for review
