+++ vignettes/using_visdat.Rmd
@@ -9,7 +9,7 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-```{r setup, echo = FALSE}
+```{r setup, echo = FALSE, include = FALSE}
 
 knitr::opts_chunk$set(fig.width = 5,
                       fig.height = 4)
