### 2017-06-24 #87 @njtierney
Hi @noamross !

I'm working on a response here - https://github.com/njtierney/visdat/blob/fix-functionality/paper/response.Rmd

Still need to wrap it all up, but I am nearly done I think!

I'm about to fly over to Europe, and give a talk/tutorial, I am going to try and do the following as quickly as possible:

- complete my review
- Submit visdat to CRAN

I will then update the visdat CRAN version once it has been onboarded into rOpenSci and describe future work.

So sorry I have been behind on this, it has been a hectic couple of months, but I really value this process, so thank you all @seaaan and @batpigandme for your patience!
