### 2017-07-02 18a3016
added an internal function to add a palette for vis_dat and vis_guess
