### 2017-07-02 2c7a63f
added palette to vis_guess and vis_dat
