+++ man/vis_guess.Rd
@@ -4,10 +4,12 @@
 \alias{vis_guess}
 \title{Visualise type guess in a data.frame}
 \usage{
-vis_guess(x)
+vis_guess(x, palette = "default")
 }
 \arguments{
 \item{x}{a data.frame object}
+
+\item{palette}{character "default", "qual" or "cb_safe". "default" (the default) provides the stock ggplot scale for separating the colours. "qual" uses an experimental qualitative colour scheme for providing distinct colours for each Type. "cb_safe" is a set of colours that are appropriate for those with colourblindness. "qual" and "cb_safe" are drawn from http://colorbrewer2.org/.}
 }
 \value{
 \code{ggplot2} object displaying the guess of the type of values in the data frame and the position of any missing values.
