### 2017-07-02 f8b21d5
improved testing with vdiffr
