### 2017-07-02 ccf802a
fixed documentation spacing at 80 characters  + indenting
next time I'll use the tidyverse linter, somehow
