+++ R/vis_guess.R
@@ -1,11 +1,21 @@
 #' Visualise type guess in a data.frame
 #'
-#' `vis_guess` visualises the class of every single individual cell in a dataframe and displays it as ggplot object, similar to `vis_dat`. Cells are coloured according to what class they are and whether the values are missing. `vis_guess` estimates the class of individual elements using `readr::guess_parser`.  Currently it may be slow on larger datasets.
+#' `vis_guess` visualises the class of every single individual cell in a
+#'   dataframe and displays it as ggplot object, similar to `vis_dat`. Cells
+#'   are coloured according to what class they are and whether the values are
+#'   missing. `vis_guess` estimates the class of individual elements using
+#'   `readr::guess_parser`.  Currently it may be slow on larger datasets.
 #'
 #' @param x a data.frame object
-#' @param palette character "default", "qual" or "cb_safe". "default" (the default) provides the stock ggplot scale for separating the colours. "qual" uses an experimental qualitative colour scheme for providing distinct colours for each Type. "cb_safe" is a set of colours that are appropriate for those with colourblindness. "qual" and "cb_safe" are drawn from http://colorbrewer2.org/.
+#' @param palette character "default", "qual" or "cb_safe". "default" (the
+#'   default) provides the stock ggplot scale for separating the colours.
+#'   "qual" uses an experimental qualitative colour scheme for providing
+#'   distinct colours for each Type. "cb_safe" is a set of colours that are
+#'   appropriate for those with colourblindness. "qual" and "cb_safe" are drawn
+#'   from http://colorbrewer2.org/.
 #'
-#' @return `ggplot2` object displaying the guess of the type of values in the data frame and the position of any missing values.
+#' @return `ggplot2` object displaying the guess of the type of values in the
+#'   data frame and the position of any missing values.
 #'
 #' @seealso [vis_miss()] [vis_dat()] [vis_miss_ly()] [vis_compare()]
 #'
