+++ R/miss_guide_label.R
@@ -4,7 +4,9 @@
 #'
 #' @param x is a dataframe passed from vis_miss(x).
 #'
-#' @return a data_frame with two columns `p_miss_lab` and `p_pres_lab`, containing the labels to use for present and missing. A dataframe is returned because I think it is a good style habit compared to a list.
+#' @return a data_frame with two columns `p_miss_lab` and `p_pres_lab`,
+#'   containing the labels to use for present and missing. A dataframe is
+#'   returned because I think it is a good style habit compared to a list.
 #'
 #'
 miss_guide_label <- function(x) {
