+++ R/vis_dat.R
@@ -1,14 +1,24 @@
 #' Visualises a data.frame to tell you what it contains.
 #'
-#' `vis_dat` gives you an at-a-glance ggplot object of what is inside a dataframe. Cells are coloured according to what class they are and whether the values are missing. As `vis_dat` returns a ggplot object, it is very easy to customize and change labels, and customize the plot
+#' `vis_dat` gives you an at-a-glance ggplot object of what is inside a
+#'   dataframe. Cells are coloured according to what class they are and whether
+#'   the values are missing. As `vis_dat` returns a ggplot object, it is very
+#'   easy to customize and change labels, and customize the plot
 #'
 #' @param x a data.frame object
 #'
-#' @param sort_type logical TRUE/FALSE. When TRUE (default), it sorts by the type in the column to make it easier to see what is in the data
+#' @param sort_type logical TRUE/FALSE. When TRUE (default), it sorts by the
+#'   type in the column to make it easier to see what is in the data
 #'
-#' @param palette character "default", "qual" or "cb_safe". "default" (the default) provides the stock ggplot scale for separating the colours. "qual" uses an experimental qualitative colour scheme for providing distinct colours for each Type. "cb_safe" is a set of colours that are appropriate for those with colourblindness. "qual" and "cb_safe" are drawn from http://colorbrewer2.org/.
+#' @param palette character "default", "qual" or "cb_safe". "default" (the
+#'   default) provides the stock ggplot scale for separating the colours.
+#'   "qual" uses an experimental qualitative colour scheme for providing
+#'   distinct colours for each Type. "cb_safe" is a set of colours that are
+#'   appropriate for those with colourblindness. "qual" and "cb_safe" are drawn
+#'   from http://colorbrewer2.org/.
 #'
-#' @return `ggplot2` object displaying the type of values in the data frame and the position of any missing values.
+#' @return `ggplot2` object displaying the type of values in the data frame and
+#'   the position of any missing values.
 #'
 #' @seealso [vis_miss()] [vis_miss_ly()] [vis_guess()] [vis_compare()]
 #'
