+++ R/visdat-package.r
@@ -1,6 +1,8 @@
 #' visdat
 #'
-#' visdat is a package that helps with the preliminary visualisation of data. visdat makes it easy to visualise your whole dataset so that you can visually identify problems.
+#' visdat is a package that helps with the preliminary visualisation of data.
+#'   visdat makes it easy to visualise your whole dataset so that you can
+#'   visually identify problems.
 #'
 #' @seealso
 #'
@@ -13,7 +15,7 @@
 #'   \item [vis_miss_ly()]
 #' }
 #'
-#' To learn more about visdat, see the vignette at \url{www.njtierney.com/visdat/articles/using_visdat.html}
+#' Learn more about visdat at \url{www.njtierney.com/visdat/articles/using_visdat.html}
 #' @name visdat
 #' @docType package
 #' @importFrom magrittr %>%
