### 2017-07-03 0f35622
minor doco changes
