+++ R/typical_data.r
@@ -20,21 +20,4 @@
 #'   \item{Died}{Whether or not this person has died yet., see ?wakefield::died}
 #' }
 #'
-#' @note the following code was used to create this data
-#' library(wakefield)
-#'
-#' set.seed(1145)
-#'
-#' typical_data <- r_data_frame(n = 5000,
-#'                              id,
-#'                              race,
-#'                              age,
-#'                              sex,
-#'                              height_cm,
-#'                              iq,
-#'                              smokes,
-#'                              income,
-#'                              died)
-#'
-#'
 "typical_data"
