+++ man/typical_data.Rd
@@ -27,21 +27,4 @@ A dataset containing information about some randomly generated people,
 created using the excellent \code{wakefield} package. It is created as
 deliberately messy dataset.
 }
-\note{
-the following code was used to create this data
-library(wakefield)
-
-set.seed(1145)
-
-typical_data <- r_data_frame(n = 5000,
-id,
-race,
-age,
-sex,
-height_cm,
-iq,
-smokes,
-income,
-died)
-}
 \keyword{datasets}
