+++ man/guess_type.Rd
@@ -17,7 +17,7 @@ an integer, "20.11" is a double, "text" is character, etc.
 \code{guess_type} is used by vis_guess to guess cell elements, like \code{fingerprint}.
 }
 \examples{
-
+\dontrun{
 guess_type(1)
 
 guess_type("x")
@@ -25,5 +25,5 @@ guess_type("x")
 guess_type(c("1", "0L"))
 
 purrr::map_df(iris, guess_type)
-
+}
 }
