### 2017-07-03 92dec1d
add prototype vis_compare_ly and vis_guess_ly and vis_dat_ly
