+++ man/vis_dat_ly.Rd
@@ -1,0 +1,37 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/vis_dat_ly.R
+\name{vis_dat_ly}
+\alias{vis_dat_ly}
+\title{Interactive plot of data in plotly}
+\usage{
+vis_dat_ly(x, sort_type = TRUE, palette = "default")
+}
+\arguments{
+\item{x}{a data.frame object}
+
+\item{sort_type}{logical TRUE/FALSE. When TRUE (default), it sorts by the
+type in the column to make it easier to see what is in the data}
+
+\item{palette}{character "default", "qual" or "cb_safe". "default" (the
+default) provides the stock ggplot scale for separating the colours.
+"qual" uses an experimental qualitative colour scheme for providing
+distinct colours for each Type. "cb_safe" is a set of colours that are
+appropriate for those with colourblindness. "qual" and "cb_safe" are drawn
+from http://colorbrewer2.org/.}
+}
+\value{
+\code{plotly} interactive plot, similar in appearance to \code{vis_dat}
+}
+\description{
+This makes \code{vis_dat} interactive using plotly. At the moment this is an
+initial implementation that just wraps \code{plotly::ggplotly()} around
+\code{vis_dat}.
+}
+\examples{
+
+vis_dat_ly(airquality)
+
+}
+\seealso{
+\code{\link[=vis_dat]{vis_dat()}} \code{\link[=vis_miss]{vis_miss()}} \code{\link[=vis_miss_ly]{vis_miss_ly()}} \code{\link[=vis_guess]{vis_guess()}} \code{\link[=vis_compare]{vis_compare()}}
+}
