### 2017-07-03 b5eaa1d
added _ly() funs to experimental vignette
