### 2017-07-03 e425ac1
use typical_data not airquality
