+++ tests/testthat/test-vis-dat.R
@@ -1,10 +1,10 @@
 context("vis_dat")
 
 # try out all the options
-vis_dat_plot <- vis_dat(airquality)
-vis_dat_plot_sort_type <- vis_dat(airquality, sort_type = FALSE)
-vis_dat_plot_pal_qual <- vis_dat(airquality, palette = "qual")
-vis_dat_plot_pal_cb <- vis_dat(airquality, palette = "cb_safe")
+vis_dat_plot <- vis_dat(typical_data)
+vis_dat_plot_sort_type <- vis_dat(typical_data, sort_type = FALSE)
+vis_dat_plot_pal_qual <- vis_dat(typical_data, palette = "qual")
+vis_dat_plot_pal_cb <- vis_dat(typical_data, palette = "cb_safe")
 
 test_that("vis_dat creates the right plot",{
   vdiffr::expect_doppelganger("vis_dat vanilla",
@@ -19,5 +19,5 @@ test_that("vis_dat creates the right plot",{
 
 
 test_that("vis_dat fails when the wrong palette is provided",{
-  testthat::expect_error(vis_dat(airquality, palette = "wat"))
+  testthat::expect_error(vis_dat(typical_data, palette = "wat"))
 })
