+++ tests/testthat/test-vis-guess.R
@@ -1,10 +1,10 @@
 context("vis_guess")
 
 # try out all the options
-vis_guess_plot <- vis_guess(airquality)
-vis_guess_plot_cb_safe <- vis_guess(airquality, palette = "cb_safe")
-vis_guess_plot_qual <- vis_guess(airquality, palette = "qual")
-vis_guess_plot_default <- vis_guess(airquality, palette = "default")
+vis_guess_plot <- vis_guess(typical_data)
+vis_guess_plot_cb_safe <- vis_guess(typical_data, palette = "cb_safe")
+vis_guess_plot_qual <- vis_guess(typical_data, palette = "qual")
+vis_guess_plot_default <- vis_guess(typical_data, palette = "default")
 
 test_that("vis_guess creates the right plot",{
   vdiffr::expect_doppelganger("vis_guess vanilla", vis_guess_plot)
@@ -14,5 +14,5 @@ test_that("vis_guess creates the right plot",{
 })
 
 test_that("vis_guess fails when the wrong palette is provided",{
-  testthat::expect_error(vis_guess(airquality, palette = "wat"))
+  testthat::expect_error(vis_guess(typical_data, palette = "wat"))
 })
