### 2017-07-03 753e12e
unexport guess_type
