### 2017-07-03 d32bd97
about 95% done on the rOpenSci review
