+++ paper/response.Rmd
@@ -36,7 +36,7 @@ Below I respond to suggested changes and other constructive criticisms.
 
 > In the "Using visdat" vignette, it says "missing data represented by black", but it shows up as gray on my computer.
 
-- I think I've fixed this now, does it still happen?
+- This is fixed this now.
 
 > I don't know why, but in the plotly section, a vis_dat() (non-interactive) plot of df_test appears between the first two interactive plots. I can't explain it, hopefully it's just a weird quirk on my computer.
 
@@ -54,7 +54,7 @@ Below I respond to suggested changes and other constructive criticisms.
 
 > Did you mean to export guess_type()?
 
-- Yes, as I thought that it might be a useful function for users, although I can't think of a good usecase outside of `vis_guess` right now. Perhaps it might be best for `guess_type` to be unexported?
+- Yes, as I thought that it might be a useful function for users, although I can't think of a good usecase outside of `vis_guess` right now. For the moment I will unexport guess_type now.
 
 > Formatting: links, code, and other formatting need to be done with .Rd syntax. For example, for code, use \code{}, not backticks. For bold, use \strong{} instead of asterisks.
 
@@ -130,11 +130,11 @@ Do you think this is OK?
 
 > The flip argument should be available for all of the functions if you provide it for any, for the sake of consistency. Possibly also the palette argument.
 
-- I agree that the rows should be in the order that you see the dataframes. I would also prefer for the column names to be on top of the visualisation - I have now implemented this, although I think that the presentation of the column names is perhaps not ahs nice.
+- I agree that the rows should be in the order that you see the dataframes. I would also prefer for the column names to be on top of the visualisation - I have now implemented this, although I think that the presentation of the column names is perhaps not as nice.
 
 - The flip argument was placed there to experiment with this idea of flipping the plot - as you suggested, to have row 1 at the top. The flip argument has now been removed, as I cannot now really think of any times where I would prefer the old arrangement of rows.
 
-- I think that the default sort_type option should be to put all of the similar columns together - to me this makes it easier to identify what is similar, and what is different; when I start looking at a dataset I usually want to know what variables are what class. I am willing to change this, but would be interested to see what others think.
+- I think that the default sort_type option should be to put all of the similar columns together - to me this makes it easier to identify what is similar, and what is different; when I start looking at a dataset I usually want to know what variables are what class. I am willing to change this, however.^
 
 - I agree with you re the palette argument being provided for all plots, although I'm not sure how this could work for vis_miss. But perhaps I should allow for users to insert their own pallete for this. Personally I think that there should be a standard palette where the common classes (character, factor, integer, numeric, datetime, etc) each have their own distinct colour, and other classes are given different factors.
 
@@ -179,28 +179,45 @@ Do you think this is OK?
 
 > vis_compare and vis_guess are indicated as being in beta, which seems also to apply to the plotly versions of the functions. The message that they emit is a helpful indication that they may change in the future. Before submitting to CRAN, however, you might consider moving those functions to the development version of the package and only uploading the functions with a stable API, then adding the beta version later once they stabilize. This is a judgment call; I think I tend towards being conservative on this issue personally.
 
-- Agreed, I need to create a dev branch, where these experimental features can reside.
+- Agreed, I will create a dev branch, where these experimental features can reside.
 
[…]
