### 2017-07-03 9725e0b
final commit for changes in https://github.com/ropensci/onboarding/issues/87
