+++ data-raw/create-typical-data-large.R
@@ -1,0 +1,6 @@
+# the following code was used to create this data
+library(wakefield)
+set.seed(1214)
+typical_data_large <- r_data_theme(n = 300, data_theme = "the_works")
+
+use_data(typical_data_large)
