### 2017-07-03 db38848
cutting down the package to minimal features for CRAN 0.1.0
