+++ paper/paper.Rmd
@@ -1,74 +1,0 @@
----
-title: 'visdat: Visualising whole data frames'
-tags:
-  - visualisation
-  - R
-  - exporatory data analysis
-authors:
- - name: Nicholas Tierney
-   orcid: 0000-0003-1460-8722
-   affiliation: 1
-affiliations:
- - name: Queensland University of Technology
-   index: 1
-date: 09 January 2017
-bibliography: paper.bib
-output: 
-  html_document: 
-    keep_md: yes
----
-
-# Summary
-
-When you get a new data set, you need to look at the data to get a sense of what it contains and potential problems with it. You read in your data, but you then need to look at it to understand what you need to do to make it ready for analysis. A phrase that often arises in this process is "looking at the data". It is helpful to look more closely at what this means. You can look at the first six rows of data, the head of the data:
-
-```{r head-iris}
-
-head(iris)
-
-```
-
-Another alterantive is `glimpse`, from the `dplyr` package [@dplyr]
-
-```{r glimpse}
-
-dplyr::glimpse(iris)
-
-```
-
-This has a better printing method that gives us more information about the data, displaying doubles and a factor. However, we don't always have data like the canonical iris dataset. Let's take a look at some data that might be a bit more typical of "messy" data.
-
-```{r visdat-glimpse}
-
-library(visdat)
-dplyr::glimpse(typical_data)
-
-```
-
-Looking at this, you might then ask:
-
-> Isn't it odd that Income is a factor? And Age is a character? 
-
-And you might start to wonder what else is different, what else changed? 
-
-And it might be a bit unclear where to go from there. Do you plot the data? Why does my plot look wierd? What are these other strange features in the data? The `visdat` package provides visualisations of an entire dataframe at once. Initially inspired by [`csv-fingerprint`](https://github.com/setosa/csv-fingerprint), `visdat` provides tools to create heatmap-like visualisations of an entire dataframe. `visdat` is an R [@Rcore] package provides 2 main functions `vis_dat` and `vis_miss`.
-
-`vis_dat()` helps explore the data class structure and missingness:
-
-```{r load-data}
-
-vis_dat(typical_data)
-
-```
-
-And the `vis_miss` function provides a custom plot for missing data.
-
-```{r vis-miss}
-
-vis_miss(typical_data)
-
-```
-
-These functions provide useful tools to help "get a look at the data", using principled visualisation techniques. The plots are built using ggplot2 [@ggplot2], which provides a consistent and powerful framework for visualisations. This means that users can customise and extend graphics from visdat very easily.
-
-# References
