+++ paper/paper.bib
@@ -1,30 +1,0 @@
-@Manual{Rcore,
-  title = {R: A Language and Environment for Statistical Computing},
-  author = {{R Core Team}},
-  organization = {R Foundation for Statistical Computing},
-  address = {Vienna, Austria},
-  year = {2016},
-  url = {https://www.R-project.org/},
-}
-
-@Manual{ggplot2,
-  title = {ggplot2: An Implementation of the Grammar of Graphics},
-  author = {Hadley Wickham and Winston Chang and RStudio},
-  year = {2016},
-  note = {R package version 2.1.0},
-  url = {https://CRAN.R-project.org/package=ggplot2},
-}
-
-@book{wickham2016,
-  title={R for Data Science},
-  author={Wickham, Hadley and Grolemund, Garrett},
-  year={2016},
-  publisher={O'Reilly Media, Inc.}
-}
-
-@Manual{dplyr,
-    title = {dplyr: A Grammar of Data Manipulation},
-    author = {Hadley Wickham and Romain Francois},
-    note = {R package version 0.5.0},
-    url = {https://cran.r-project.org/web/packages/dplyr/},
-  }
