+++ R/vis_miss.R
@@ -22,7 +22,7 @@
 #' @return `ggplot2` object displaying the position of missing values in the
 #'   dataframe, and the percentage of values missing and present.
 #'
-#' @seealso [vis_miss_ly()] [vis_dat()] [vis_guess()] [vis_compare()]
+#' @seealso [vis_dat()]
 #'
 #' @examples
 #'
