+++ R/vis_dat.R
@@ -20,7 +20,7 @@
 #' @return `ggplot2` object displaying the type of values in the data frame and
 #'   the position of any missing values.
 #'
-#' @seealso [vis_miss()] [vis_miss_ly()] [vis_guess()] [vis_compare()]
+#' @seealso [vis_miss()]
 #'
 #' @examples
 #'
