### 2017-07-03 7f74fc9
update description and cran comments
