+++ cran-comments.md
@@ -7,10 +7,6 @@
 0 errors | 0 warnings | 0 notes
 There were no ERRORs or WARNINGs or NOTEs
 
-There was 1 NOTE:
-
-* This is a new release.
-
 ## Reverse dependencies
 
 This is a new release, so there are no reverse dependencies.
