### 2017-07-04 #87 @arfon
> These references are provided in the yaml, and in a .bib file - hopefully these get compiled with the JOSS paper perhaps @arfon can chime in here?

👍 there should be nothing under the `# References` section in the `paper.md` file as they will be compiled in by pandoc (see an example paper here https://github.com/arfon/fidgit/blob/master/paper/paper.md)
