### 2017-07-04 2c63b05
move readme figures into man/
