+++ README.md
@@ -1,10 +1,11 @@
+
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 visdat
 ======
 
 <!-- add a TravisCI badge -->
 <!-- Add an appVeyor badge -->
-[![Travis-CI Build Status](https://travis-ci.org/njtierney/visdat.svg?branch=master)](https://travis-ci.org/njtierney/visdat) [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/njtierney/visdat?branch=master&svg=true)](https://ci.appveyor.com/project/njtierney/visdat)[![Coverage Status](https://img.shields.io/codecov/c/github/njtierney/visdat/master.svg)](https://codecov.io/github/njtierney/visdat?branch=master)
+[![Travis Build Status](https://travis-ci.org/njtierney/visdat.svg?branch=master)](https://travis-ci.org/njtierney/visdat)[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/njtierney/visdat?branch=master&svg=true)](https://ci.appveyor.com/project/njtierney/visdat)[![Coverage Status](https://img.shields.io/codecov/c/github/njtierney/visdat/master.svg)](https://codecov.io/github/njtierney/visdat?branch=master)
 
 How to install
 ==============
@@ -47,7 +48,7 @@ library(visdat)
 vis_dat(airquality)
 ```
 
-![](README-figs/README-vis-dat-aq-1.png)
+![](man/figures/README-vis-dat-aq-1.png)
 
 The plot above tells us that R reads this dataset as having numeric and integer values, with some missing data in `Ozone` and `Solar.R`. The classes are represented on the legend, and missing data represented by grey. The column/variable names are listed on the x axis.
 
@@ -59,7 +60,7 @@ vis_dat(airquality,
         sort_type = FALSE)
 ```
 
-![](README-figs/README-vis-dat-aq-sort-type-1.png)
+![](man/figures/README-vis-dat-aq-sort-type-1.png)
 
 ### With many kinds of data
 
@@ -70,7 +71,7 @@ To demonstrate what visdat looks like when you have different kinds of data, we 
 vis_dat(typical_data)
 ```
 
-![](README-figs/README-vis-dat-typical-data-1.png)
+![](man/figures/README-vis-dat-typical-data-1.png)
 
 We can also look into using even wider data, looking at `typical_larger_data`
 
@@ -79,7 +80,7 @@ We can also look into using even wider data, looking at `typical_larger_data`
 vis_dat(typical_data_large)
 ```
 
-![](README-figs/README-vis-dat-typical-larger-data-1.png)
+![](man/figures/README-vis-dat-typical-larger-data-1.png)
 
 Using `vis_miss()`
 ------------------
@@ -91,7 +92,7 @@ We can explore the missing data further using `vis_miss()`.
 vis_miss(airquality)
 ```
 
-![](README-figs/README-vis-miss-aq-1.png)
+![](man/figures/README-vis-miss-aq-1.png)
 
 The percentages of missing/complete in `vis_miss` are accurate to 1 decimal place.
 
@@ -103,7 +104,7 @@ vis_miss(airquality,
          cluster = TRUE)
 ```
 
-![](README-figs/README-vis-miss-aq-cluster-1.png)
+![](man/figures/README-vis-miss-aq-cluster-1.png)
 
 The columns can also just be arranged by columns with most missingness, by setting `sort_miss = TRUE`.
 
@@ -113,7 +114,7 @@ vis_miss(airquality,
          sort_miss = TRUE)
 ```
 
[…]
