### 2017-07-04 a8af88a
no longer need covr or vdiffr from github
