+++ .travis.yml
@@ -4,10 +4,6 @@ language: R
 sudo: false
 cache: packages
 
-r_github_packages:
-  - jimhester/covr
-  - lionel-/vdiffr
-
 after_success:
   - Rscript -e 'library(covr); codecov()'
 
