### 2017-07-05 e0c4cc9
update DESCRIPTION
