+++ DESCRIPTION
@@ -1,9 +1,9 @@
 Package: visdat
-Title: Preliminary Visualisation of Data
+Title: Preliminary Data Visualisation
 Version: 0.1.0
 Authors@R: person("Nicholas", "Tierney", email = "nicholas.tierney@gmail.com", role = c("aut", "cre"))
-Description: Preliminary exploratory data visualisation of an entire dataset to 
-  identify problems or unexpected features.
+Description: Create preliminary exploratory data visualisations of an entire 
+  dataset to identify problems or unexpected features using 'ggplot2'.
 Depends:
     R (>= 3.2.2)
 License: MIT + file LICENSE
