+++ tests/testthat/test-vismiss-legend.R
@@ -1,7 +1,6 @@
 ## testing for miss_guide_label
 context("vis_miss Legend")
 
-
 test_that("miss_guide_label works for exactly 0.1% missing", {
 
   # 1 missing in 1000.
