### 2017-07-05 #87 @batpigandme
@njtierney the mods look great.
Q. 1: I like the Beta warning, personally, though I'm sure this is a different strokes for different folks matter.
Q. 2: I prefer the visual analogue to the data frame, but, again that's totally stylistic.
Q. 3: Yes, this is great. I don't mind if things take a while, but I love having a heads up if it's gonna happen.
Q. 5: The `pandoc-citeproc` issue was, indeed, about my machine, and now that that's fixed it's running as it should.
