### 2017-07-05 #87 @seaaan
The changes look good after a quick look. I should be able to respond fully
to the modifications this weekend.
