### 2017-07-08 #87 @seaaan
# Summary
I respond below to a few specific things, but in general for the sake of brevity I didn't respond when I agreed with your changes. So mentally insert "awesome change! you rock!" everywhere that I didn't respond :)

The two points of substance I have, as you'll see, are:

* Please implement the warning for large data sets with `vis_dat` (to save we low RAM types from ourselves). This is the only change that needs to be done in my opinion before acceptance and approval.
* For speeding up `vis_guess` I do think a pull request to `readr` is the way to go. I think it's fine to approve the package in the meantime and have that as a future issue to work on.

# Details
Did you mean to export guess_type()?

- Yes, as I thought that it might be a useful function for users, although I can't think of a good usecase outside of vis_guess right now. For the moment I will unexport guess_type now.

```
To clarify I'm not opposed to exporting it. What I meant to draw attention to (but
didn't actually say!) was that its documentation makes it sound like an internal
function not intended for public use. If exported, the documentation should be changed.
```

Formatting: links, code, and other formatting need to be done with .Rd syntax. For example, for code, use \code{}, not backticks. For bold, use \strong{} instead of asterisks.

- I have now converted this over to use markdown with roxygen, this should be fixed now.

```
Based on Noam's comment, I was actually in the wrong here. Sorry about that!
```

The paper doesn't have any references in the "References" section, but does have inline references.

- These references are provided in the yaml, and in a .bib file - hopefully these get compiled with the JOSS paper perhaps @arfon can chime in here?

```
Sorry, my bad!
```

I agree that the rows should be in the order that you see the dataframes. I would also prefer for the column names to be on top of the visualisation

- I have now implemented this, although I think that the presentation of the column names is perhaps not as nice.

```
Just a minor comment here, for the new plots with the column names on top, I don't
think you  need the title "variables in dataset."
```

This `[vis_compare]` is a neat function! It would be cool if you could generalize it to also handle data frames of unequal dimensions, perhaps by showing which rows/columns are only in one data frame or the other. Just an idea I had though, not a requirement at all!

- I am glad you like this! It is a difficult problem that I cannot work out, re the rows/cols matching when they are out of the dimension. I'm not quite sure what to do here to make this possible, I think that this should be tackled in a future release of visdat, and will be moved to a development branch.

```
Fair! I agree that it's a difficult problem. My thought was something along the lines of
 displaying a df the size of the bigger of the two dfs and indicating which rows and
columns are present in the bigger and not in the smaller. That of course leaves the
 problem of some rows/cols missing in one df and some others missing in the other df,
where you need to somehow display the union of the two. Though perhaps at this level
the use case on such dissimilar dfs breaks down.
```

`readr` stuff for efficiency in `vis_guess`.
 - OK, so I think that this is a fantastic idea! But I'm not sure on the best way to move this into production of visdat, since it sounds like we would need to borrow a chunk of code from readr and then modify it. I wonder if perhaps the solution here is to do a pull request for this specific feature in readr?

```
I think yeah asking them to export a vectorised collectorGuess function from readr
would be the way to go. It should be relatively straight-forward from what I did last
time to submit a pull request if they're open to it. It's possible that a vectorised
function actually already exists in readr and I just missed it -- there's lots of code
in that package!

If they don't want to export such a function, you could include a bunch of readr code
in visdat but it would be a lot of code because the parsing functions are complicated.
```
# Answers to your questions
I agree with @batpigandme's answers to your questions.

Q. 3: I agree with Mara. Please do implement the warning and save me from myself: My dinky computer runs out of memory and dies trying to display large data frames.
Q. 4: Answered above.
Q. 5: Happily, the plotly issue is no more.

Great job on the changes!
