### 2017-07-09 #87 @noamross
Thanks for continued great progress, everyone!  Let me just chime in with that it's fine to defer a more efficient `vis_guess`to a future version if it first requires updates to **readr**.  Also I'll put in a request for a `warn_large_data` argument, maybe with a default of `interactive()`, so that users can skip the prompt for big data frames when scripting.
