### 2017-07-10 #87 @njtierney
@batpigandme Thank you again for your comments! :)

Q. 2: I prefer the visual analogue to the data frame, but, again that's totally stylistic.

> OK, I think I will go with my default of `sort_type == TRUE`, but this can still be changed by the user, so this should be OK. I might change this in the future though, I imagine as I teach more I might change my mind!
