### 2017-07-10 317508d
update CRAN comments ... prior to rhub
