### 2017-07-10 a4b8509
add print for freetype version as per https://github.com/lionel-/vdiffr/issues/19#issuecomment-314103083
