+++ tests/testthat/test-vis-miss.R
@@ -7,6 +7,8 @@ vis_miss_plot_sort_rows <- vis_miss(typical_data, sort_miss = TRUE)
 vis_miss_plot_show_perc <- vis_miss(typical_data, show_perc = FALSE)
 
 test_that("vis_miss creates the right plot",{
+  ver <- as.character(gdtools::version_freetype())
+  cat(sprintf("FreeType version: %s\n", ver))
   vdiffr::expect_doppelganger("vis_miss vanilla", vis_miss_plot)
   vdiffr::expect_doppelganger("vis_miss cluster", vis_miss_plot_cluster)
   vdiffr::expect_doppelganger("vis_miss sort rows", vis_miss_plot_sort_rows)
