### 2017-07-10 a8d43e5
remove x axis label - no longer needed
