### 2017-07-11 a63d475
add skip_on_cran() for vdiffr tests
