+++ tests/testthat/test-vis-dat.R
@@ -7,6 +7,7 @@ vis_dat_plot_pal_qual <- vis_dat(typical_data, palette = "qual")
 vis_dat_plot_pal_cb <- vis_dat(typical_data, palette = "cb_safe")
 
 test_that("vis_dat creates the right plot",{
+  skip_on_cran()
   ver <- as.character(gdtools::version_freetype())
   cat(sprintf("FreeType version: %s\n", ver))
   vdiffr::expect_doppelganger("vis_dat vanilla",
