### 2017-07-11 4d30d91
add test for the large visualisation problem
