### 2017-07-11 #87 @njtierney
@seaaan  Thank you again for your comments! :D Likewise, please insert "Awesome! Thank you!" For each of these :)

Re `guess_type` - I think that this function, whilst kinda useful, doesn't really belong to the user, so I think the move to make it internal is better, and makes the purpose of the package, pre-exploratory visualisation, clearer.

Re `x axis label of "variables in dataset"` - I agree, this has been changed in a recent push to the cran branch, great point!

Re `vis_compare`, I need to think about this some more, for the moment, perhaps we can label this as an upcoming feature?

Re `vectorized guess`, perhaps we can post an issue on `readr` and ask the tidyverse team?

@noamross I have just added the `warn_large_data` param for both `vis_dat` and `vis_miss` - which you can see [here](https://github.com/njtierney/visdat/blob/cran-0-1-0/R/vis_miss.R#L45)

Just to clarify - I am working on getting a cut down, preliminary version of visdat onto CRAN that just has `vis_dat` and `vis_miss`, then we can go from there, does that work for everyone?
