### 2017-07-11 aa39b75
final CRAN comments
