+++ cran-comments.md
@@ -4,12 +4,13 @@
 * win-builder (devel and release)
 
 ## R CMD check results
-0 errors | 0 warnings | 1 notes
+0 errors | 0 warnings | 1 note
+
 There were no ERRORs or WARNINGs
 
 There was 1 note
 
-* This is a new release
+* This is a new submission
 
 ## Reverse dependencies
 
