### 2017-07-11 5c56656
update cran comments
