### 2017-07-24 c12b1b7
bump to 0.1.0.9000 + add cran comments
