+++ cran-comments.md
@@ -1,0 +1,18 @@
+## Test environments
+* local OS X install, R 3.4.1
+* ubuntu 12.04 (on travis-ci), R 3.4.1
+* win-builder (devel and release)
+
+## R CMD check results
+0 errors | 0 warnings | 1 note
+
+There were no ERRORs or WARNINGs
+
+There was 1 note
+
+* Maintainer: 'Nicholas Tierney <nicholas.tierney@gmail.com>'
+* This is a new submission
+
+## Reverse dependencies
+
+This is a new release, so there are no reverse dependencies.
