### 2017-07-24 6bee194
rbuildignore + travis + DESCRIPTION
