+++ .travis.yml
@@ -1,19 +1,24 @@
 # R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
 
 language: R
-sudo: false
 cache: packages
+sudo: true
+dist: trusty
 
-r_github_packages:
-  - jimhester/covr
-  - lionel-/vdiffr
+r:
+#  - oldrel
+  - release
+  - devel
+
+env:
+  global:
+  - _R_CHECK_FORCE_SUGGESTS_=false
 
 after_success:
-  - Rscript -e 'library(covr); codecov()'
+  - Rscript -e 'covr::codecov()'
 
-addons:
-  apt:
-    sources:
-      - debian-sid
-    packages:
-      - libfreetype6
+before_install:
+  - sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable --yes
+  - sudo apt-get --yes --force-yes update -qq
+  - sudo apt-get install --yes libudunits2-dev libproj-dev libgeos-dev libgdal-dev
+  - Rscript -e 'update.packages(ask = FALSE)'
