### 2017-07-24 877269f
misc CRAN updates
