+++ R/vis_dat.R
@@ -1,49 +1,59 @@
 #' Visualises a data.frame to tell you what it contains.
 #'
-#' `vis_dat` gives you an at-a-glance ggplot object of what is inside a dataframe. Cells are coloured according to what class they are and whether the values are missing. As `vis_dat` returns a ggplot object, it is very easy to customize and change labels, and customize the plot
+#' `vis_dat` gives you an at-a-glance ggplot object of what is inside a
+#'   dataframe. Cells are coloured according to what class they are and whether
+#'   the values are missing. As `vis_dat` returns a ggplot object, it is very
+#'   easy to customize and change labels, and customize the plot
 #'
 #' @param x a data.frame object
 #'
-#' @param sort_type logical TRUE/FALSE. When TRUE (default), it sorts by the type in the column to make it easier to see what is in the data
+#' @param sort_type logical TRUE/FALSE. When TRUE (default), it sorts by the
+#'   type in the column to make it easier to see what is in the data
 #'
-#' @param palette character "default", "qual" or "cb_safe". "default" (the default) provides the stock ggplot scale for separating the colours. "qual" uses an experimental qualitative colour scheme for providing distinct colours for each Type. "cb_safe" is a set of colours that are appropriate for those with colourblindness. "qual" and "cb_safe" are drawn from http://colorbrewer2.org/.
+#' @param palette character "default", "qual" or "cb_safe". "default" (the
+#'   default) provides the stock ggplot scale for separating the colours.
+#'   "qual" uses an experimental qualitative colour scheme for providing
+#'   distinct colours for each Type. "cb_safe" is a set of colours that are
+#'   appropriate for those with colourblindness. "qual" and "cb_safe" are drawn
+#'   from http://colorbrewer2.org/.
 #'
-#' @param flip logical TRUE, will flip the axis labels to be on top, resembling a dataframe
+#' @param warn_large_data logical default is TRUE
 #'
-#' @return `ggplot2` object displaying the type of values in the data frame and the position of any missing values.
+#' @param large_data_size integer default is 900000, this can be changed.
 #'
-#' @seealso [vis_miss()] [vis_miss_ly()] [vis_guess()] [vis_compare()]
+#' @return `ggplot2` object displaying the type of values in the data frame and
+#'   the position of any missing values.
+#'
+#' @seealso [vis_miss()]
 #'
 #' @examples
 #'
 #' vis_dat(airquality)
 #'
-#' # flip the axis to look more like a dataframe
-#' vis_dat(airquality, flip = TRUE)
-#'
 #' # experimental colourblind safe palette
 #' vis_dat(airquality, palette = "cb_safe")
+#' vis_dat(airquality, palette = "qual")
 #'
 #' @export
 vis_dat <- function(x,
                     sort_type = TRUE,
                     palette = "default",
-                    flip = FALSE) {
+                    warn_large_data = TRUE,
+                    large_data_size = 900000) {
 
-  # x  = airquality
-  if (sort_type == TRUE) {
+  # add warning for large data
+  if (ncol(x) * nrow(x) > large_data_size && warn_large_data){
+    stop("Data exceeds recommended size for visualisation, please consider
+         downsampling your data, or set argument 'warn_large_data' to FALSE.")
+  }
 
-    # arrange by the columns with the highest missingness
-    # code inspired from
-    # https://r-forge.r-project.org/scm/viewvc.php/pkg/R/...
-    # ...missing.pattern.plot.R?view=markup&root=mi-dev
-    # get the order of columns with highest missingness
+  if (sort_type) {
 
     type_sort <- order(
       # get the class, if there are multiple classes, combine them together
       purrr::map_chr(.x = x,
[…]
