+++ R/vis_miss.R
@@ -1,20 +1,32 @@
 #' Visualise a data.frame to display missingness.
 #'
-#' `vis_miss` provides an at-a-glance ggplot of the missingness inside a dataframe, colouring cells according to missingness, where black indicates a missing cell and grey indicates a present cell. As it returns a ggplot object, it is very easy to customize and change labels, and so on.
+#' `vis_miss` provides an at-a-glance ggplot of the missingness inside a
+#'   dataframe, colouring cells according to missingness, where black indicates
+#'   a missing cell and grey indicates a present cell. As it returns a ggplot
+#'   object, it is very easy to customize and change labels.
 #'
 #' @param x a data.frame
 #'
-#' @param cluster logical TRUE/FALSE. TRUE specifies that you want to use hierarchical clustering (mcquitty method) to arrange rows according to missingness. FALSE specifies that you want to leave it as is
+#' @param cluster logical. TRUE specifies that you want to use hierarchical
+#'   clustering (mcquitty method) to arrange rows according to missingness.
+#'   FALSE specifies that you want to leave it as is.
 #'
-#' @param sort_miss logical TRUE/FALSE. TRUE arranges the columns in order of missingness
+#' @param sort_miss logical. TRUE arranges the columns in order of missingness
 #'
-#' @param show_perc logical TRUE/FALSE. TRUE now adds in the \% of missing/complete data in the whole dataset into the legend. Default value is TRUE
+#' @param show_perc logical. TRUE now adds in the \% of missing/complete data
+#'   in the whole dataset into the legend. Default value is TRUE.
 #'
-#' @param flip logical if TRUE, will flip the axis labels to be on top, resembling a dataframe
+#'@param show_perc_col logical. TRUE adds in the \% missing data in a given
+#'  column into the x axis. Can be disabled with FALSE
 #'
-#' @return `ggplot2` object displaying the position of missing values in the dataframe, and the percentage of values missing and present.
+#' @param warn_large_data logical default is TRUE
 #'
-#' @seealso [vis_miss_ly()] [vis_dat()] [vis_guess()] [vis_compare()]
+#' @param large_data_size integer default is 900000, this can be changed.
+#'
+#' @return `ggplot2` object displaying the position of missing values in the
+#'   dataframe, and the percentage of values missing and present.
+#'
+#' @seealso [vis_dat()]
 #'
 #' @examples
 #'
@@ -24,22 +36,27 @@
 #'
 #' vis_miss(airquality, sort_miss = TRUE)
 #'
-#' # flip the axis to look more like a dataframe
-#' vis_miss(airquality, flip = TRUE)
-#'
 #' @export
 vis_miss <- function(x,
                      cluster = FALSE,
                      sort_miss = FALSE,
                      show_perc = TRUE,
-                     flip = FALSE){
+                     show_perc_col = TRUE,
+                     large_data_size = 900000,
+                     warn_large_data = TRUE){
+
+  # add warning for large data
+  if (ncol(x) * nrow(x) > large_data_size && warn_large_data) {
+    stop("Data exceeds recommended size for visualisation, please consider
+         downsampling your data, or set argument 'warn_large_data' to FALSE.")
+  }
+
   # make a TRUE/FALSE matrix of the data.
   # This tells us whether it is missing (true) or not (false)
-  # x = airquality
   x.na <- is.na(x)
 
   # switch for creating the missing clustering
-  if (cluster == TRUE){
+  if (cluster){
 
     # this retrieves a row order of the clustered missingness
     row_order_index <-
@@ -52,7 +69,7 @@ vis_miss <- function(x,
     row_order_index <- seq_len(nrow(x))
   } # end else
 
-  if (sort_miss == TRUE) {
[…]
