### 2017-07-24 b6e93b6
update pkg docs
