+++ man/vis_dat.Rd
@@ -4,12 +4,8 @@
 \alias{vis_dat}
 \title{Visualises a data.frame to tell you what it contains.}
 \usage{
-<<<<<<< HEAD
 vis_dat(x, sort_type = TRUE, palette = "default", warn_large_data = TRUE,
   large_data_size = 9e+05)
-=======
-vis_dat(x, sort_type = TRUE, palette = "default")
->>>>>>> master
 }
 \arguments{
 \item{x}{a data.frame object}
@@ -23,13 +19,10 @@ default) provides the stock ggplot scale for separating the colours.
 distinct colours for each Type. "cb_safe" is a set of colours that are
 appropriate for those with colourblindness. "qual" and "cb_safe" are drawn
 from http://colorbrewer2.org/.}
-<<<<<<< HEAD
 
 \item{warn_large_data}{logical default is TRUE}
 
 \item{large_data_size}{integer default is 900000, this can be changed.}
-=======
->>>>>>> master
 }
 \value{
 \code{ggplot2} object displaying the type of values in the data frame and
