### 2017-07-24 c9f8a13
rerun build_site and set eval = FALSE for experimental features vignette
