+++ R/visdat-package.r
@@ -22,4 +22,4 @@
 NULL
 
 if(getRversion() >= "2.15.1")  utils::globalVariables(c("."))
-globalVariables(c("valueGuess"))
+globalVariables(c("valueGuess", "valueType"))
