### 2017-07-24 f9b8617
add test for removing the missing %s from the columns
