+++ tests/testthat/test-vis-miss.R
@@ -5,6 +5,7 @@ vis_miss_plot <- vis_miss(typical_data)
 vis_miss_plot_cluster <- vis_miss(typical_data, cluster = TRUE)
 vis_miss_plot_sort_rows <- vis_miss(typical_data, sort_miss = TRUE)
 vis_miss_plot_show_perc <- vis_miss(typical_data, show_perc = FALSE)
+vis_miss_plot_show_perc_col <- vis_miss(typical_data, show_perc_col = FALSE)
 
 test_that("vis_miss creates the right plot",{
   skip_on_cran()
@@ -14,4 +15,5 @@ test_that("vis_miss creates the right plot",{
   vdiffr::expect_doppelganger("vis_miss cluster", vis_miss_plot_cluster)
   vdiffr::expect_doppelganger("vis_miss sort rows", vis_miss_plot_sort_rows)
   vdiffr::expect_doppelganger("vis_miss show percent", vis_miss_plot_show_perc)
+  vdiffr::expect_doppelganger("vis_miss show percent in columns", vis_miss_plot_show_perc_col)
 })
