### 2017-07-25 1762682
explicitly state else if ~sort_miss
