### 2017-07-28 b315382
Fix links to vignettes
Note only fixed in Rmd, not compiled
