### 2017-07-28 #87 @njtierney
Thanks @noamross !

Important question - how do I best add you all into the DESCRIPTION file?
