### 2017-07-28 #87 @noamross
We're not done until we're done @njtierney 😉

We are currently working with CRAN to get a `rev` (reviewer) role approved for the `Authors@R` field - it has been accepted in the past but inconsistently.  Past authors have, _if reviewers agree and authors desire it_, added reviewers as `ctb` contributors with some extra comments in the metadata.  See here for an example: https://github.com/ropensci/roadoi/blob/master/DESCRIPTION As editor I should not be included in the authors field.
