### 2017-07-28 #87 @njtierney
OK, sure thing, that makes sense!

Would you mind if I added you all to the `Thank Yous` section at the end of visdat?
