### 2017-07-28 #87 @seaaan
Do you mean to be exporting `label_col_missing_pct`? It is exported but the
documentation states that it is internal. Otherwise, looks good to me! I've
been using the package when I get new data sets lately and it's really
helpful for getting a quick overview of problems.

I'd be happy to be included in the thank yous section :)

On Thu, Jul 27, 2017 at 10:12 PM, Nicholas Tierney <notifications@github.com
> wrote:

> OK, sure thing, that makes sense!
>
> Would you mind if I added you all to the Thank Yous section at the end of
> visdat?
>
