### 2017-07-28 67f2dee
remove label_col_missing_pct as described in https://github.com/ropensci/onboarding/issues/87
