+++ NAMESPACE
@@ -1,6 +1,5 @@
 # Generated by roxygen2: do not edit by hand
 
-export(label_col_missing_pct)
 export(vis_compare)
 export(vis_compare_ly)
 export(vis_dat)
