### 2017-07-28 cd8723b
added acknowledgement of Noam Ross, Sean Huges, and Mara Averick (https://github.com/ropensci/onboarding/issues/87)
