+++ README.Rmd
@@ -151,4 +151,6 @@ Thank you to Hadley Wickham for suggesting the use of the internals of `readr` t
 
 Thank you to Miles McBain for his suggestions on how to improve `vis_guess`. This resulted in making it at least 2-3 times faster.
 
-Thanks also to Carson Sievert for writing the code that combined `plotly` with `visdat`, and for Noam Ross for suggesting this in the first place.
+Thanks to Carson Sievert for writing the code that combined `plotly` with `visdat`, and for Noam Ross for suggesting this in the first place.
+
+Finally thank you to [rOpenSci](https://github.com/ropensci) and it's amazing [onboarding process](https://github.com/ropensci/onboarding), this process has made visdat a much better package, thanks to the editor Noam Ross (@noamross), and the reviewers Sean Hughes (@seaaan) and Mara Averick (@batpigandme).
