### 2017-07-28 #87 @njtierney
Oh, thanks for picking up on that! I've just removed it now and added y'all in the Thank yous in the referenced commits above :D

Lemme know what you all think.
