### 2017-07-28 #87 @batpigandme
Other than a renegade missing _s_ (just submitted [PR](https://github.com/njtierney/visdat/pull/52)), it all looks great. @njtierney and `visdat` FTW!
