### 2017-07-29 20f8add
update description URL, close #50
