### 2017-07-31 #87 @noamross
Approved! Thanks @njtierney for submitting and @batpigandme and @seaaan for your reviews and all of you hanging on for the long haul!

To-dos:
```
[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
```
- [ ] Fix any links in badges for CI and coverage to point to the ropensci URL. (We'll turn on the services on our end as needed)
- [ ] Submit to JOSS using the Zenodo DOI.  We will tag it for expedited review.

Welcome aboard! I believe you already have plans for a blog post to follow this up.
