### 2017-08-01 43f9583
add note that visdat is on CRAN, add the rOpenSci footer :tada:
