### 2017-08-01 #87 @njtierney
Hooray! 🎉

1. visdat has been transferred, I've joined the visdat team
2. I've added the rOpenSci footer
3. It looks like the badges are working,

Just working on activating Zenodo and touching up the JOSS paper, will report back ASAP
