+++ data-raw/create-typical-larger-data.R
@@ -1,6 +1,0 @@
-# the following code was used to create this data
-library(wakefield)
-set.seed(1214)
-typical_larger_data <- r_data_theme(n = 300, data_theme = "the_works")
-
-use_data(typical_larger_data)
