+++ data-raw/create-typical-data.R
@@ -11,5 +11,10 @@ typical_data <- r_data_frame(n = 5000,
                             iq,
                             smokes,
                             income,
-                            died)
-use_data(typical_data)
+                            died) %>%
+  wakefield::r_na(cols = c(2,3,6,8),
+                  prob = 0.1) %>%
+mutate(Income = as.factor(Income),
+       Age = as.character(Age))
+
+use_data(typical_data, overwrite = TRUE)
