### 2017-08-01 24c11b0
doco typo for vis_compare
