### 2017-08-01 4d391b1
minor changes to paper - add a sentence to vis_dat and vis_miss, and change style slightly to be more appropriate for a journal.
