+++ paper/paper.bib
@@ -24,7 +24,8 @@
 
 @Manual{dplyr,
     title = {dplyr: A Grammar of Data Manipulation},
-    author = {Hadley Wickham and Romain Francois},
-    note = {R package version 0.5.0},
-    url = {https://cran.r-project.org/web/packages/dplyr/},
+    author = {Hadley Wickham and Romain Francois and Lionel Henry and Kirill Müller},
+    year = {2017},
+    note = {R package version 0.7.2},
+    url = {https://CRAN.R-project.org/package=dplyr},
   }
