+++ paper/paper.Rmd
@@ -9,9 +9,9 @@ authors:
    orcid: 0000-0003-1460-8722
    affiliation: 1
 affiliations:
- - name: Queensland University of Technology
+ - name: Monash University
    index: 1
-date: 09 January 2017
+date: 01 August 2017
 bibliography: paper.bib
 output: 
   html_document: 
@@ -20,7 +20,7 @@ output:
 
 # Summary
 
-When you get a new data set, you need to look at the data to get a sense of what it contains and potential problems with it. You read in your data, but you then need to look at it to understand what you need to do to make it ready for analysis. A phrase that often arises in this process is "looking at the data". It is helpful to look more closely at what this means. You can look at the first six rows of data, the head of the data:
+Reading in a new dataset means looking at the data to get a sense of what it contains, and potential problems and challenges with the data to get it analysis ready. "Looking at the data" can mean different things. Often times you look at the first six rows of data, the head of the data:
 
 ```{r head-iris}
 
@@ -28,7 +28,7 @@ head(iris)
 
 ```
 
-Another alterantive is `glimpse`, from the `dplyr` package [@dplyr]
+Or you can `glimpse` the data, using the `dplyr` package [@dplyr]
 
 ```{r glimpse}
 
@@ -36,7 +36,7 @@ dplyr::glimpse(iris)
 
 ```
 
-This has a better printing method that gives us more information about the data, displaying doubles and a factor. However, we don't always have data like the canonical iris dataset. Let's take a look at some data that might be a bit more typical of "messy" data.
+This shows that there are 150 observations, and 5 variables, which are doubles and a factor, along with some of the values in the data. However, we usually don't have data like the canonical iris dataset. Let's take a look at some data that might be a bit more typical of "messy" data.
 
 ```{r visdat-glimpse}
 
@@ -45,15 +45,11 @@ dplyr::glimpse(typical_data)
 
 ```
 
-Looking at this, you might then ask:
+Looking at this, you might then ask: "Isn't it odd that Income is a factor? And Age is a character?". And you might start to wonder what else is different, what else changed? It might be a bit unclear where to go from there. Do you do some exploratory plots of the data? What if the plot looks weird because the data has strange types? What if the data has other strange features? 
 
-> Isn't it odd that Income is a factor? And Age is a character? 
+There is a need for a tool to do preliminary data visualisation, to identify these problems early. The `visdat` package provides this, creating visualisations of an entire dataframe at once. Initially inspired by [`csv-fingerprint`](https://github.com/setosa/csv-fingerprint), `visdat` provides tools to create heatmap-like visualisations of an entire dataframe. `visdat` is an R [@Rcore] package provides 2 main functions `vis_dat` and `vis_miss`.
 
-And you might start to wonder what else is different, what else changed? 
-
-And it might be a bit unclear where to go from there. Do you plot the data? Why does my plot look wierd? What are these other strange features in the data? The `visdat` package provides visualisations of an entire dataframe at once. Initially inspired by [`csv-fingerprint`](https://github.com/setosa/csv-fingerprint), `visdat` provides tools to create heatmap-like visualisations of an entire dataframe. `visdat` is an R [@Rcore] package provides 2 main functions `vis_dat` and `vis_miss`.
-
-`vis_dat()` helps explore the data class structure and missingness:
+`vis_dat()` helps explore the data class structure and missingness, by displaying the class for each variable and the missing data, and presenting the plot in an intuitive way - it reads top down as the data would. The columns are grouped by similar class as well, to put similarly classified data types together.
 
 ```{r load-data}
 
@@ -61,7 +57,7 @@ vis_dat(typical_data)
 
 ```
 
-And the `vis_miss` function provides a custom plot for missing data.
[…]
