### 2017-08-01 #87 @noamross
Reviewers may be added with the annotated `ctb` as in the role to DESCRIPTION (as in the *roadoi* example), but don't list editors.
