### 2017-08-01 4fd3c7b
add Mara and Sean to the DESCRIPTION as `ctb`.
