### 2017-08-01 9def41d
removed blank lines from DESCRIPTION
