+++ DESCRIPTION
@@ -10,8 +10,6 @@ Authors@R: c(
     reviewed the package for rOpenSci, 
     see https://github.com/ropensci/onboarding/issues/87")
     )
-
-
 Description: visdat makes it easy to visualise your whole dataset so that you
     can visually identify problems.
 Depends:
