### 2017-08-01 feb3c99
update the pkgdown docs
