### 2017-08-01 01ff3a5
Add DOI badge :badger emoji:
