### 2017-08-01 #87 @njtierney
Alrighty!

4. Activated Zenodo watching the repo
5. tag and create a release so as to create a zenodo version and DOI

![image](https://user-images.githubusercontent.com/6488485/28826730-79a2dc20-770e-11e7-81e7-ee3de58ffdbe.png)

I'm not sure what to do about the badges for appveyor and codecov, can I help with that?

Working on submitting visdat to JOSS now, I might do this tomorrow morning as it's getting a bit late 😪
