### 2017-08-01 #87 @njtierney
OK...I think I'm done! Except I submitted to JOSS using the github repo not the Zenodo repo link: https://zenodo.org/record/837274 Oops!
